% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_stat.R
\name{CO_stat}
\alias{CO_stat}
\title{Cross-over statistics}
\usage{
CO_stat(mppData)
}
\arguments{
\item{mppData}{An object of class \code{mppData}}
}
\value{
\code{data.frame} with the number of cross-over per cross, average
number of cross-over per cross per genotype, and the average number of
cross-over per genotype per chromosome.
}
\description{
Function that counts the total number of cross-over (CO) per cross, the number of cross
over per cross per genotype and the number of cross-over.
}
\examples{

# Later

}
\author{
Vincent Garin
}
