% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_ABH.R
\name{cross_ABH}
\alias{cross_ABH}
\title{ABH assignment per cross}
\usage{
cross_ABH(par.sc, off.sc, cross.ind, par.per.cross)
}
\arguments{
\item{par.sc}{\code{Character} marker scores \code{matrix} of the
parents of the different cross. \strong{The rownames
must be the parents identifiers and correspond to the one used in argument
\code{par.par.cross}. Parent scores must be homozygous.
Missing value must be coded NA.}}

\item{off.sc}{\code{Character} marker scores \code{matrix} of the
offspring. \strong{The possible values must be: homozygous like parent 1 or 2,
heterozygous or missing. Missing values must be coded NA.}}

\item{cross.ind}{\code{Character} vector with the same length as the number
of offspring genotypes which specifies to which cross each offspring genotype
belong.}

\item{par.per.cross}{Three columns \code{Character matrix} specifying :
1) the cross indicators (\strong{In the same order as they appear in
\code{cross.ind}}); 2) the parents 1 identifiers of the crosses;
3) the parents 2 identifiers of the crosses. \strong{The cross indicators
must be  similar to the one used in the arguement \code{cross.ind}.
The list of parent identifiers must be the same to the rownames of
the argument \code{par.sc}.}}
}
\value{
Return:

\item{geno.ABH}{\code{Character matrix} with ABH coding for the different
genotypes of the population}
}
\description{
Attributes ABH or NA (missing) code to the offspring according to the scores
of parents 1 and 2 of each cross.
}
\details{
The function takes successively the parents of the different cross as
reference and assign the following scores: "A" if the offspring score is
equivalent to parent 1; "B" if it is equivalent to parent 2; "H" if it is
heterozygous. The function attributes NA for missing when: 1) the offspring
score is missing; 2) the two parents have the same score; or
3) when at least one parental score is missing.
}
\examples{

# Genotype and phenotype dataset
data("USNAM_geno")
geno <- USNAM_geno

# Remove marker for which parents are monomorphic
par.mono <- QC_MAF(mk.mat = geno[1:6, ])
geno <- geno[, -which(par.mono == 0)]

# Remove marker for which at least one parent is heterozygous
par.het <- QC_hetero(mk.mat = geno[1:6, ])
geno <- geno[, -which(par.het != 0)]

# Parents scores
par.sc <- geno[1:6, ]

# Offspring scores
off.sc <- geno[7:506, ]

# Cross.indicator
cross.ind <- substr(rownames(geno)[7:506], 1, 4)

# Parent of the crosses matrix
par.per.cross <- cbind(unique(cross.ind), rep("B73",5), rownames(par.sc)[-1])

geno.ABH <- cross_ABH(par.sc = par.sc, off.sc = off.sc, cross.ind = cross.ind,
                      par.per.cross = par.per.cross)

}
\seealso{
\code{\link{mppData_form}}, \code{\link{USNAM_genoABH}}
}
\author{
Vincent Garin
}
