% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_partition.R
\name{CV_partition}
\alias{CV_partition}
\title{Cross validation partition}
\usage{
CV_partition(cross.ind, k = 5)
}
\arguments{
\item{cross.ind}{\code{Character} vector with the same length as the number
of genotypes which specifies to which cross each genotype belong.}

\item{k}{\code{Numeric} value representing the number of subsets (fold) into
which data are spread within cross. Default = 5.}
}
\value{
Return:

\item{fold}{\code{List} of k lists (one for each fold). Each fold list contain two
vectors with genotypes indices of the training set (\code{$train.set}) and
the one of the validation set (\code{$val.set}).}
}
\description{
Makes a partition of the genotype indicators for a single MPP QTL analysis
cross-validation (CV) run. The genotype indicators are randomly assigned to
k subsets (folds). The function returns a list of k partitions containing one
vector with the indices of the training set and another with the genotypes
indices of the validation set. Each of the k partition is sucessively used as
validation set. The rest of the data going to the training set.
}
\examples{

data(USNAM_mppData)

part.cv <- CV_partition(cross.ind = USNAM_mppData$cross.ind, k = 5)

part.cv[[1]]$train.set
part.cv[[1]]$val.set

}
\seealso{
\code{\link{mpp_CV}}
}
\author{
Vincent Garin
}
