% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_connectedeness.R
\name{design_connectedness}
\alias{design_connectedness}
\title{Determines the connected parts of a design}
\usage{
design_connectedness(par.per.cross, plot.des = TRUE, output.loc = NULL)
}
\arguments{
\item{par.per.cross}{Three columns \code{character matrix} specifying :
1) the cross indicators ; 2) the parents 1 identifiers of the crosses;
3) the parents 2 identifiers of the crosses.}

\item{plot.des}{\code{Logical} value indicating if connected part of the
design should be plotted. Default = TRUE.}

\item{output.loc}{Path where the plot of the design will be saved if the
argument is given. Default = NULL.}
}
\value{
Return a list with each element representing one connected part of the design
and the list of parents contained in this part.
}
\description{
Uses the metod of Weeks and Williams (1964) and the package igraph to
determine connected parts of the design.
}
\examples{

data(USNAM_mppData)

par.per.cross <- USNAM_mppData$par.per.cross

con.part <- design_connectedness(par.per.cross)

}
\author{
Vincent Garin
}
\references{
Weeks, D. L., & Williams, D. R. (1964). A note on the determination of
connectedness in an N-way cross classification. Technometrics, 6(3), 319-324.
}

