% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_012.R
\name{geno_012}
\alias{geno_012}
\title{Marker scores into 0, 1, 2 format}
\usage{
geno_012(mk.mat)
}
\arguments{
\item{mk.mat}{\code{Character} genotype  marker score \code{matrix}.
\strong{Marker scores must be coded using one letter for each allele.
For example, AA, CC, GG, TT, AC, AG, AT, CA, CG, CT, GA, GC, GT, TA, TC, TG.
Missing values must be coded NA.}}
}
\value{
Return:

\code{List} with two matrices

\item{geno012}{Marker \code{matrix} with marker scores coded as 0, 1, 2
corresponding to the number of copies of the least frequent SNP allele.}

\item{all.ref}{\code{matrix} with reference allele scores. The first row
represents the allele with the highest MAF, the second the one with the
lowest MAF and the two others the heterozygous scores.}
}
\description{
Transform genotype marker score into 0, 1, 2 format. The score represents the
number of copies of the allele with the lowest marker allele frequency (MAF).
}
\examples{

data(USNAM_geno)

data <- geno_012(mk.mat = USNAM_geno)

data_012 <- data[[1]]

}
\seealso{
\code{\link{mppData_form}}
}
\author{
Vincent Garin
}
