% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_ABH_prob.R
\name{geno_ABH_prob}
\alias{geno_ABH_prob}
\title{Probability of the genotype}
\usage{
geno_ABH_prob(mppData)
}
\arguments{
\item{mppData}{An object of class \code{mppData}}
}
\value{
Vector of parental orignin genotype probability.
}
\description{
Function that calculate the probability of parental origin (AA), (AB) or (BB)
genotypes.
}
\examples{

# Later

}
\author{
Vincent Garin
}
