% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_par_seg.R
\name{imp_par_seg}
\alias{imp_par_seg}
\title{Impute parent marker scores}
\usage{
imp_par_seg(mppData)
}
\arguments{
\item{mppData}{An object of class \code{mppData} formed with
\code{\link{create.mppData}}.}
}
\value{
parent marker score matrix with some missing marker imputed
}
\description{
Impute the parent marker scores based on offpring segregation. The function
impute parent marker score in situation where one parent is missing and
the other is homozygous. In that case, it looks at the cross segregation.
If there is segratation, the missing parent get the allele marker score
different than other parent. If there is no segregation, the missing parent
gets the same score as the homozygous parent.
}
\author{
Vincent Garin
}
