% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_beagle.R
\name{impute_beagle}
\alias{impute_beagle}
\title{Impute maker data using Beagle}
\usage{
impute_beagle(
  mk.mat,
  map,
  alleles,
  beagle.loc,
  undiferentiate.het.score = TRUE
)
}
\arguments{
\item{mk.mat}{\code{Character} genotype  marker score \code{matrix}.
\strong{Marker scores must be coded using one letter for each allele.
For example, AA, CC, GG, TT, AC, AG, AT, CA, CG, CT, GA, GC, GT, TA, TC, TG.
Missing values must be coded NA.}}

\item{map}{Four columns \code{data.frame} with: 1) marker names, 2) chromosome,
3) position in cM, 4) physical bp position (must be in ascending order).}

\item{alleles}{4xN_marker matrix with SNP marker alleles of dimensions
4 x N_marker: homozygous score minor allele (first row), 
homozygous score major allele (second row), heterozygous score 1, heterozygous
score 2.}

\item{beagle.loc}{\code{Character} path specifying the location where the
Beagle software is located.}

\item{undiferentiate.het.score}{\code{Logical} value specifying if in the imputed
value the two different phased heterozygous scores (e.g. AT and TA) should be
considered as similar (e.g. AT). Default = TRUE.}
}
\value{
imputed genotype marke matrix
}
\description{
Impute maker data using Beagle (Browning & Browning, 2018).
}
\references{
B L Browning, Y Zhou, and S R Browning (2018). A one-penny imputed genome
from next generation reference panels. Am J Hum Genet 103(3):338-348.
doi:10.1016/j.ajhg.2018.07.015
}
\author{
Vincent Garin
}
