% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncMat_cross.R
\name{IncMat_cross}
\alias{IncMat_cross}
\title{Cross effect incidence matrix}
\usage{
IncMat_cross(cross.ind)
}
\arguments{
\item{cross.ind}{\code{Character} or \code{factor} vector specifying to
which cross the genotypes belong.}
}
\value{
Return:

\item{cross.mat}{Cross effect incidence matrix composed of 0 and 1 value
indicating to which cross each genotype belongs.}
}
\description{
Form a cross effect incidence matrix.
}
\examples{

data(USNAM_mppData)

cross.mat <- IncMat_cross(cross.ind = USNAM_mppData$cross.ind)

}
\seealso{
\code{\link{IncMat_parent}}, \code{\link{IncMat_QTL}}
}
\author{
Vincent Garin
}
