% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncMat_parent.R
\name{IncMat_parent}
\alias{IncMat_parent}
\title{Skeleton of parental incidence matrices}
\usage{
IncMat_parent(mppData)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}
}
\value{
Return:

\code{List} with two matrices

\item{PA}{Incidence matrix specifying for each genotype which parent
is the parent 1 or A}

\item{PB}{Incidence matrix specifying for each genotype which parent
is the parent 2 or B}
}
\description{
Form the skeleton matrices used to form parental incidence
matrices in parental effect models.
}
\examples{

data(USNAM_mppData)

par.mat <- IncMat_parent(USNAM_mppData)

}
\seealso{
\code{\link{IncMat_cross}}, \code{\link{IncMat_QTL}}
}
\author{
Vincent Garin
}
