% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncMat_QTL.R
\name{IncMat_QTL}
\alias{IncMat_QTL}
\title{QTL incidence matrix}
\usage{
IncMat_QTL(x, mppData, Q.eff, par.clu, cross.mat, par.mat, order.MAF = FALSE)
}
\arguments{
\item{x}{\code{Integer} value indicating the genetic position on the map
\code{mppData$map} for which the QTL incidence matrix should be returned.}

\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. for more details see \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument for the ancesral model \code{(Q.eff = "anc")}.
\code{interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{cross.mat}{Cross effect incidence matrix that can be obtained with
\code{\link{IncMat_cross}}.}

\item{par.mat}{Parents incidence matrix that can be obtained with
\code{\link{IncMat_parent}}.}

\item{order.MAF}{\code{Logical} value specifying if the QTL incidence matrix
should be ordered by allele frequency for a parental and ancestral QTL
incidence matrix. The colum will be ordred from the least to the most frequent
allele. Default = FALSE.}
}
\value{
Return:

\item{QTL.mat}{QTL incidence matrix or vector (for bi-allelic model). For the
cross-specific model, it represent the difference between the
number of allele from parent 2 or B and parent 1 or A divided by two. For
parental (ancestral) model it represent the expected number of parent
(ancestor) allele copies. For the bi-allelic model, it represent the number
of copies of the allele with the highest frequency.}
}
\description{
Builds a single position QTL incidences matrix
}
\examples{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)

cross.mat <- IncMat_cross(cross.ind = USNAM_mppData$cross.ind)
par.mat <- IncMat_parent(USNAM_mppData)
par.clu <- USNAM_parClu

QTLmatCr <- IncMat_QTL(x = 2, mppData = USNAM_mppData,
                       cross.mat = cross.mat, Q.eff = "cr")

QTLmatPar <- IncMat_QTL(x = 2, mppData = USNAM_mppData, par.mat = par.mat,
                        Q.eff = "par")

QTLmatAnc <- IncMat_QTL(x = 2, mppData = USNAM_mppData, par.mat = par.mat,
                        par.clu = par.clu, Q.eff = "anc")

QTLmatBi <- IncMat_QTL(x = 2, mppData = USNAM_mppData_bi, Q.eff = "biall")


}
\seealso{
\code{\link{IncMat_cross}}, \code{\link{IncMat_parent}}, 
\code{\link{mpp_SIM}}, \code{\link{parent_cluster}},
\code{\link{USNAM_parClu}}
}
\author{
Vincent Garin
}
