% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_012_mk_score_from_ABH.R
\name{infer_012_mk_score_from_ABH}
\alias{infer_012_mk_score_from_ABH}
\title{Infer 012 marker score from ABH information}
\usage{
infer_012_mk_score_from_ABH(mppData, geno.par)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{geno.par}{Parents genotype marker \code{matrix} with parent as row
and markers as columns. column names should be the marker identifiers
Row names should be the parents identifiers.}
}
\value{
Offspring genotype marker matrix in 012 format. 0: homozygous major allele,
1: heterozygous, 2: homozygous minor allele.
}
\description{
For each genotype marker scores, the function look at the most probable
parental origin (A, B or H), translate this information into an ATCG marker
score looking at the parent scores, and finally into 012 score. Such function
should be called if parents have non missing information. Marker imputation
of the parents can be done using \code{\link{impute_beagle}}.
}
