% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpp_proc.R
\name{mpp_proc}
\alias{mpp_proc}
\title{MPP QTL analysis}
\usage{
mpp_proc(pop.name = "MPP", trait.name = "trait1", mppData, Q.eff = "cr",
  par.clu = NULL, VCOV = "h.err", est.gen.eff = FALSE, thre.cof = 3,
  win.cof = 20, N.cim = 1, window = 20, thre.QTL = 3, win.QTL = 20,
  backward = TRUE, alpha.bk = 0.05, ref.par = NULL, CI = FALSE,
  drop = 1.5, text.size = 18, parallel = FALSE, cluster = NULL,
  verbose = TRUE, output.loc = getwd())
}
\arguments{
\item{pop.name}{\code{Character} name of the studied population.
Default = "MPP".}

\item{trait.name}{\code{Character} name of the studied trait.
Default = "trait1".}

\item{mppData}{An object of class \code{mppData}. See
\code{\link{mppData_form}} for details.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. For more details see \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument for the ancesral model \code{(Q.eff = "anc")}.
\code{interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{est.gen.eff}{\code{Logical} value. if \code{est.gen.eff = TRUE},
the function will save the decomposed genetic effects per cross/parent.
These results can be ploted with the function \code{\link{plot_genEffects}}
to visualize a genome-wide decomposition of the genetic effects.
\strong{This functionality is ony available for the cross-specific,
parental and ancestral models.}
Default value = FALSE.}

\item{thre.cof}{\code{Numeric} value representing the -log10(p-value)
threshold
above which a position can be peaked as a cofactor. Significance threshold
values can be obtained by permutation using \code{\link{mpp_perm}} function.
Default = 3.}

\item{win.cof}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected cofactors. Default = 20.}

\item{N.cim}{\code{Numeric} value specifying the number of time the CIM
analysis is repeated. Default = 1.}

\item{window}{\code{Numeric} distance on the left an right of a cofactor
position where it is not included in the model. Default = 20.}

\item{thre.QTL}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as QTL. Significance
threshold values can be obtained by permutation using \code{\link{mpp_perm}}
function. Default = 3.}

\item{win.QTL}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected QTL. Default = 20.}

\item{backward}{\code{Logical} value. If \code{backward = TRUE},
the function performs a backward elimination on the list of selected QTLs.
Default = TRUE.}

\item{alpha.bk}{\code{Numeric} value indicating the significance level for
the backward elimination. Terms with p-values above this value will
iteratively be removed. Default = 0.05.}

\item{ref.par}{Optional \code{Character} expression defining the parental
allele that will be used as reference for the parental model. For the
ancestral model, the ancestral class containing the reference parent will be
set as reference. \strong{This option can only be used if the MPP design is
composed of a unique connected part}. Default = NULL.}

\item{CI}{\code{Logical} value. If \code{CI = TRUE}, the function will
compute a -log10(pval) drop confidence interval for each QTL after
calculating a CIM- profile (without cofactors on the scanned chromosome).
Default = FALSE.}

\item{drop}{\code{Numeric} -log10(p-value) drop value at the limits of the
interval. Default = 1.5.}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 18.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run function in parallel user must
provide cluster in the \code{cluster} argument. \strong{Parallelization is
only available for HRT (linear) models \code{VCOV = "h.err"}}.
Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}

\item{verbose}{\code{Logical} value indicating if the steps of mpp_proc should
be printed. It will not affect the printing of the other functions called by
\code{mpp_proc()}, especially the printing of \code{asreml()}. Default = TRUE.}

\item{output.loc}{Path where a folder will be created to save the results.
By default the function uses the current working directory.}
}
\value{
Return:

List containing the following items:

\item{n.QTL}{Number of detected QTLs}

\item{cofactors}{\code{Data.frame} with cofactors positions.}

\item{QTL}{\code{Data.frame} with QTL positions.}

\item{R2}{\code{List} containing R squared statistics of the QTL effects.
for details see \code{\link{QTL_R2}} output section.}

\item{QTL.effects}{\code{List} of QTLs genetic effects. For details see
\code{\link{QTL_genEffects}} output section.}

\item{QTL.CI}{If \code{CI = TRUE}, confidence interval information of
the QTLs.}

Some output files are also saved at the specified location
(\code{output.loc}):

\enumerate{

\item{A QTL report (QTL_REPORT.txt) with: 1) the number of detected QTLs;
2) the global R squared statistics; 3) for each QTL, position information
(plus confidence interval if \code{CI = TRUE}) and estimated QTL genetic
effects per cross or parents (for details see \code{\link{QTL_genEffects}}).}

\item{The SIM and CIM results in a text file (SIM.txt, CIM.txt).}

\item{The list of cofactors (cofactors.txt).}

\item{The list of QTL (QTL.txt).}

\item{The QTL R squared statistics (QTL_R2.txt) (for details see
\code{\link{QTL_R2}}).}

\item{If \code{CI = TRUE}, the QTL confidence intervals (QTL_CI.txt).}

\item{General results of the QTL detection process: number of QTLs and
global adjusted and non-adjusted R squared statistics (QTL_genResults.txt).}

\item{The plot of the CIM profile (QTL_profile.pdf) with dotted vertical
lines representing the cofactors positions. If \code{est.gen.eff = TRUE},
plot of the genetic effects per cross or parents (gen_eff.pdf) with dashed
lines representing the QTL positions. For more details see
\code{\link{plot_QTLprof}} and \code{\link{plot_genEffects}}.}

}
}
\description{
Multi-parent QTL analysis based on models with different possible assumptions
concerning the number of alleles at the QTL position and the variance
covariance structure (VCOV) of the model. For more details about the different
models, see documentation of the function \code{\link{mpp_SIM}}.
}
\details{
\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.

The procedure is the following:

\enumerate{

\item{Simple interval mapping (SIM) to select cofactor
(\code{\link{mpp_SIM}}).}

\item{Composite interval mapping (CIM) with selected cofactors
(\code{\link{mpp_CIM}}).}

\item{Optional backward elimination on the list of QTL
candidates (\code{backward = TRUE}) (\code{\link{mpp_BackElim}}).}

\item{Computation of the QTL genetic effects (\code{\link{QTL_genEffects}})
and proportion of the phenotypic variation explained by the QTLs (R squared)
(\code{\link{QTL_R2}}).}

\item{Optional QTL confidence interval computation from a CIM- profile
(excluding cofactors on the scanned chromosome) (\code{argument CI=TRUE})
(\code{\link{QTL_CI}}).}

}
}
\examples{
 
\dontrun{

data(USNAM_mppData)
data(USNAM_mppData_bi)

# Detect the number of cores and build cluster

library(parallel)
n.cores <- detectCores()
cluster <- makeCluster(n.cores-1)

# Specify a location where your results will be saved
my.loc <- "C:/.../..."

# Cross-specific model

USNAM_cr <- mpp_proc(pop.name = "USNAM", trait.name = "ULA",
                     mppData = USNAM_mppData, est.gen.eff = TRUE, CI = TRUE,
                     parallel = TRUE, cluster = cluster, output.loc = my.loc)

# Bi-allelic model

USNAM_biall <- mpp_proc(pop.name = "USNAM", trait.name = "ULA",
                     mppData = USNAM_mppData_bi,Q.eff = "biall",
                     output.loc = my.loc)

}


}
\seealso{
\code{\link{mppData_form}},
\code{\link{mpp_BackElim}},
\code{\link{mpp_CIM}},
\code{\link{mpp_perm}},
\code{\link{mpp_SIM}},
\code{\link{parent_cluster}},
\code{\link{plot_genEffects}},
\code{\link{plot_QTLprof}},
\code{\link{QTL_CI}},
\code{\link{QTL_genEffects}},
\code{\link{QTL_R2}},
\code{\link{USNAM_parClu}}
}
\author{
Vincent Garin
}
