% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppData_chgPheno.R
\name{mppData_chgPheno}
\alias{mppData_chgPheno}
\title{Change phenotypic values in a mppData object}
\usage{
mppData_chgPheno(mppData, trait)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{trait}{Two columns \code{data.frame} with : 1) \code{character}
genotypes identifiers; 2) \code{numeric} trait values. \strong{The genotypes
identifiers must be identical to \code{mppData$geno.id}.}}
}
\value{
Return:

\item{mppData}{New \code{mppData} object with replaced phenotypic values.}
}
\description{
Replaces the phenotypic values of a \code{mppData} object by new trait values.
}
\examples{

data(USNAM_mppData)
data(USNAM_pheno)

trait <- USNAM_pheno

mppData <- mppData_chgPheno(mppData = USNAM_mppData, trait = trait)

}
\seealso{
\code{\link{mppData_form}}, \code{\link{mppData_subset}},
\code{\link{mppData_mdfPedigree}}
}
\author{
Vincent Garin
}
