% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppData_form.R
\name{mppData_form}
\alias{mppData_form}
\title{Data for MPP QTL analysis}
\usage{
mppData_form(geno.off, geno.par = NULL, IBS = FALSE,
  IBS.format = "ATCG", type, F.gen = NULL, BC.gen = NULL,
  type.mating = NULL, map, trait, cross.ind, par.per.cross,
  step = 10000, error.prob = 1e-04, map.function = "haldane",
  stepwidth = "max", dir = getwd())
}
\arguments{
\item{geno.off}{\code{Character} or \code{Numeric matrix} representing
genotype marker scores with genotypes as row and markers as column. For
cross-specific, parental and ancestral models (\code{IBS = FALSE}) marker
data must be in ABH format (for details see \code{\link{cross_ABH}}).

For bi-allelic models (\code{IBS = TRUE}), the marker scores in
\code{geno.off} can be coded using letters using one letter for each allele.
For example, AA, CC, GG, TT, AC, AG, AT, CA, CG, CT, GA, GC, GT, TA, TC, TG
(\code{IBS.format = "ATCG"}). \code{geno.off} marker score can also be
already coded in 0, 1, 2 format (\code{IBS.format = "012"}). 

For both type of mppData object (bi-allelic or not), \strong{the row names
must be the genotypes identifiers similars to the one of
trait argument. The column names must be the marker
identifiers similar to the first column of the map. Missing value
must be coded \code{NA}.}}

\item{geno.par}{Optional argument used as indication in the results for the
QTL genetic effects. \code{geno.par} is a \code{character matrix}
representing the marker scores of the parents with genotypes as row and
markers as column. \strong{The column names must be similar as the marker
list in \code{map} and \code{geno.off} arguments. The rownames must represent
the parents identifiers and be identical to the parent list given in
\code{par.per.cross}}. If \code{IBS.format = "012"}, \code{geno.par} is
set to NULL. Default = NULL.}

\item{IBS}{\code{Logical} value. Put \code{IBS = TRUE} if you want to
obtain an IBS \code{mppData} object for the bi-allelic model.
Default = FALSE.}

\item{IBS.format}{\code{Character} indicator for the format of \code{geno.off}
marker score for an IBS \code{mppData} object. \code{IBS.format = "ATCG"}, if
marker scores are coded with one letter per allele. \code{IBS.format = "012"},
if marker scores are coded in 0, 1, 2 format. Default = "ATCG".}

\item{type}{\code{Character} indicator for the type of population analysed:
type = "F" for Fn (F cross n gernerations); type = "BC" for BCn (backcross
n generations); type = "BCsFt" for backcross followed by selfing;
type = "DH" for double haploids; and type = "RIL"
for recombinant inbred lines. For RIL type specify if the population was
obtain using selfing or sibling mating using \code{type.mating}.
If type = "RIL" or "DH", heterozygous marker scores of the geno.off argument
must be set as missing (NA).}

\item{F.gen}{\code{Numeric} integer representing the number of F generations.
For example F.gen = 2 for F2. Default = NULL.}

\item{BC.gen}{\code{Numeric} integer representing the number of
backcross generations. For example BC.gen = 1 for single backcross.
Default = NULL.}

\item{type.mating}{\code{Character} specifying for a RIL population if it was
obtained by selfing ("selfing") or by sibling mating ("sib.mat").
Default = NULL.}

\item{map}{Three columns \code{data.frame} with: 1) marker or in between
position identifiers; 2) chromosome; 3) positions in centi-Morgan.
\strong{The marker identifiers must be identical to the column names of
\code{geno.off}}.}

\item{trait}{Two columns \code{data.frame} with : 1) \code{character}
genotypes identifiers; 2) \code{numeric} trait values. \strong{The genotypes
identifiers must be identical to the rownames of \code{geno.off}).}}

\item{cross.ind}{\code{Character} vector with the same length as the number of
genotypes which specifies to which cross each genotype belongs.}

\item{par.per.cross}{Three columns \code{Character matrix} specifying :
1) the cross indicators (\strong{The cross indicators must be  similar to
the one used in \code{cross.ind} and appear in the same order}); 2) the
parents 1 identifiers of the crosses; 3) the parents 2 identifiers of the
crosses. \strong{The list of parent identifiers must be similar to the
rownames of \code{geno.par}}.}

\item{step}{\code{Numeric} value representing the maximum distance (in cM)
between positions at which the genotype probabilities are calculated.
Default value = 10000 (the function do not add position in between markers).}

\item{error.prob}{\code{Numeric} value for assumed genotyping error rate
used in the calculation of the penetrance Pr(observed genotype | true genotype).
Default = 0.0001.}

\item{map.function}{\code{Character} expression specifying the type of map
function used to infer the IBD probabilities. possibility to choose
between "haldane", "kosambi","c-f","morgan".
Parameter for the function \code{calc.genoprob()}. Default = "haldane".}

\item{stepwidth}{\code{Character} expression indicating if the intermediate
points should be fixed ("fixed") or if the algorithm should insert a minimal
number of points in between positions ("max"). If stepwidth = "max" the
maximum distance between points is step. Default = "max".}

\item{dir}{Path where a .csv file Cross_object.csv will be stored to be
used for cross object formation. By default, the function uses the current
working directory.}
}
\value{
Return: a {list} of class \code{mppData} containing the following items :

\item{geno}{For a cross-specific, parental and ancestral model
(\code{IBS = FALSE}), cross object obtained with R/qtl functions
The main elements of the cross object are the IBD probabilities. For a
bi-allelic model (\code{IBS = TRUE}), marker score matrix recoded
as 0, 1, 2 according to the copies number of the most frequent allele.}

\item{allele.ref}{If \code{IBS = TRUE}, \code{matrix} with reference allele
scores. The first row represents the minor allele (lowest frequency),
the second the one represent the major allele (largest frequency) and the
two others the heterozygous scores.}

\item{geno.id}{ \code{Character} vector of genotpes identifiers.}

\item{geno.par}{Genotypic map and parent genotype marker matrix if argument
\code{geno.par} was provided by the user. For a cross-specific, parental
and ancestral model, if positions in between markers were added during
the IBD computation, these position are filled as "-".}

\item{map}{\code{Data.frame} with four columns: 1) Marker or in between
position names; 2) chromosomes; 3) position numbers on the chromosome;
4) positions in cM.}

\item{trait}{\code{Data.frame} with phenotypic values.}

\item{cross.ind}{\code{Character} vector specifying to which cross each
genotype belongs.}

\item{ped.mat}{Four columns \code{data.frame}: 1) the type of genotype:
"offspring" for the last genration and "founder" for the genotypes above
the offspring in the pedigree; 2) the genotype indicator; 3-4) the parent 1
(2) of each line.}

\item{par.per.cross}{Same object as the one given by the user in
\code{par.per.cross}.} 

\item{parents}{\code{Character} vector of parents.}  

\item{n.cr}{Number of crosses.}

\item{n.par}{Number of parents.}

\item{type}{\code{Character} expression indicating the type of population.}

\item{n.zigo}{\code{Numeric} value Indicating the number of differenct
genotypes: 2 (AA/BB) or 3 (AA/AB/BB)}

\item{biall}{\code{Logical} value specifying if data are made for a
bi-allelic model}
}
\description{
Form a single data object to perform MPP QTL anlyses.
}
\details{
\code{mppData_form} combines into a single data object the all the necessary
data to run QTL analyses except the parent clustering results
(\code{\link{parent_cluster}}). Two types of \code{mppData} objects can be
produced. The first one (\code{IBS = FALSE}), for cross-specific, parental and
ancestral models calculates identical by descent (IBD) probabilities using
\code{read.cross()} and \code{calc.genoprob()} functions from the R/qtl package
(Broman et al. 2009). F-type (F), back-cross (BC), double
haploid (DH) and recombinant imbred lines (RIL) are the allowed population type.
DH and RIL populations are read as back-cross by R/qtl. For these two
population types, no heterozygosity is allowed. It must be converted into missing.
\code{\link{USNAM_mppData}} is an example of an IBD \code{mppData} object.

The second type of \code{mppData} objects (\code{IBS = TRUE}) uses identical
by state (IBS) genetic predictors for bi-allelic model. In such case,
genotype marker score are transformed into 0, 1, 2 format representing
the copy number of the least frequent allele. For IBS \code{mppData} objects,
the user can provide imputed data for the argument \code{geno.off}. Marker
score imputation can be done using \code{\link{QC_proc}}. The markers scores
can be provided to \code{geno.off} using letters (1 letter per allele)
setting (\code{IBS.format = "ATCG"}) or already in 0, 1, 2 format
(\code{IBS.format = "012"}). \code{\link{USNAM_mppData_bi}} is
an example of an IBS \code{mppData} object.
}
\examples{
 
# Mpp data for cross, parental or ancestral model
#################################################

# Genotypes
data(USNAM_geno)
data(USNAM_genoABH)
# For details about ABH data transformation see cross_ABH() examples

geno.par <- USNAM_geno[1:6, colnames(USNAM_genoABH)]

# Phenotypes
data(USNAM_pheno)

trait <- data.frame(rownames(USNAM_pheno), USNAM_pheno[, 1],
                   stringsAsFactors = FALSE)
colnames(trait) <- c('genotypes', 'ULA')
rownames(trait) <- rownames(USNAM_pheno)

# Map
data(USNAM_map)

# Same list of markers as in the genotype matrix
map <- USNAM_map[USNAM_map[, 1] \%in\% colnames(USNAM_genoABH),]

cross.ind <- substr(rownames(USNAM_pheno), 1, 4)

par.per.cross <- cbind(unique(cross.ind), rep("B73", 5), rownames(geno.par)[-1])

\dontrun{

# Specify path to a directory
my.dir <- "C:/.../"

# IBD mppData for cross, parental or ancestral model
data <- mppData_form(geno.off = USNAM_genoABH, geno.par = geno.par,
                     IBS = FALSE, type = "F", F.gen = 6, map = map,
                     trait = trait, cross.ind = cross.ind,
                     par.per.cross = par.per.cross, step = 5,
                     map.function = "haldane",  stepwidth = "max",
                     dir = my.dir)
}


# Mpp data for bi-allelic model
###############################

# Genotypes
data(USNAM_geno)

geno <- USNAM_geno[, QC_MAF(USNAM_geno) != 0] # remove monomorphic positions
geno.par <- geno[1:6, ]

# Phenotypes
data(USNAM_pheno)

trait <- data.frame(rownames(USNAM_pheno), USNAM_pheno[, 1],
                   stringsAsFactors = FALSE)
colnames(trait) <- c('genotypes', 'ULA')
rownames(trait) <- rownames(USNAM_pheno)

# Map
data(USNAM_map)

# Same list of markers as in the genotype matrix
map <- USNAM_map[USNAM_map[, 1] \%in\% colnames(geno),]

cross.ind <- substr(USNAM_pheno[, 1], 1, 4)

par.per.cross <- cbind(unique(cross.ind), rep("B73", 5), rownames(geno.par)[-1])


\dontrun{

# Specify path to a directory
my.dir <- "C:/.../"

# IBS mppData for bi-allelic model
data_IBS <- mppData_form(geno.off = geno[7:506,], geno.par = geno.par,
                           type = "F", F.gen = 6, IBS = TRUE,
                           IBS.format = "ATCG", map = map,
                           trait = trait, cross.ind = cross.ind,
                           par.per.cross = par.per.cross, dir = my.dir)

}

}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. Bioinformatics 19:889-890.

Broman, K. W., & Sen, S. (2009). A Guide to QTL Mapping with R/qtl (Vol. 46).
New York: Springer.
}
\seealso{
\code{\link{cross_ABH}}, \code{\link{mppData_subset}},
\code{\link{pedigree_update.mppData}},
\code{\link{mpp_SIM}}, \code{\link{parent_cluster}},
\code{\link{USNAM_mppData}}, \code{\link{USNAM_mppData_bi}}
}
\author{
Vincent Garin
}
