\name{mppData_init}
\alias{mppData_init}
\docType{data}
\title{
  \code{mppData} object with raw data
}
\description{
  
  \code{mppData} object with raw genotypic and phenotypic data of a sample of the maize US nested association mapping (NAM) population. Different operations
  of quality control and data processing still need to be performed before
  the QTL detection analysis.
}
\usage{data(mppData_init)}
\format{
  \code{mppData}
}
\details{
  see examples of the \code{\link{create.mppData}}.
}

\examples{
  data(mppData_init)
}

\seealso{
  \code{\link{create.mppData}}
}

\keyword{datasets}