% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppData_subset.R
\name{mppData_subset}
\alias{mppData_subset}
\title{Subset mppData object}
\usage{
mppData_subset(mppData, mk.list = NULL, gen.list = NULL)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{mk.list}{\code{Character} vector, \code{numeric} position vector or
\code{logical} vector representing marker or in between positions that must
be subseted. Default = NULL.}

\item{gen.list}{\code{Character} vector, \code{numeric} position vector or
\code{logical} vector representing genotypes that must be subseted.
Default = NULL.}
}
\value{
Return:

\item{mppData_red}{Subseted \code{mppData} object with selected markers
or selected genotypes.}
}
\description{
Subset \code{mppData} objects by markers (\code{mk.list}) or by genotypes
(\code{gen.list}).
}
\examples{

### Marker subset

data(USNAM_mppData)

# Random selection of markers
mk.list <-  sample(USNAM_mppData$map[, 1], 50)
mppData_sub <- mppData_subset(mppData = USNAM_mppData, mk.list = mk.list)

# Selection of chromosome 1 marker
mk.list <-  (USNAM_mppData$map[, 2] == 1)
mppData_sub <- mppData_subset(mppData = USNAM_mppData, mk.list = mk.list)

### Genotype subset

# Random selection of genotypes
gen.list <-  sample(USNAM_mppData$geno.id, 200)
mppData_sub <- mppData_subset(mppData = USNAM_mppData, gen.list = gen.list)

# Selection of genotype from cross 2 and 5
crosses <- unique(USNAM_mppData$cross.ind)
gen.list <-  USNAM_mppData$geno.id[USNAM_mppData$cross.ind \%in\% crosses[c(2, 5)]]
mppData_sub <- mppData_subset(mppData = USNAM_mppData, gen.list = gen.list)

}
\seealso{
\code{\link{mppData_form}}
}
\author{
Vincent Garin
}
