% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQE_CIM.R
\name{MQE_CIM}
\alias{MQE_CIM}
\title{Composite interval maping for MQE model}
\usage{
MQE_CIM(mppData = NULL, mppData_bi = NULL, Q.eff = "cr", par.clu = NULL,
  VCOV = "h.err", cofactors = NULL, cof.Qeff, chg.Qeff = FALSE,
  window = 20, parallel = FALSE, cluster = NULL)
}
\arguments{
\item{mppData}{An IBD object of class \code{mppData}
See \code{\link{mppData_form}} for details. Default = NULL.}

\item{mppData_bi}{Required IBS object of class \code{mppData} if the user
wants to allow QTLs with a bi-allelic effect. \strong{The list of marker must
be strictly the same as the one of \code{mppData}.} Default = NULL.}

\item{Q.eff}{\code{Character} expression indicating the type of QTL effect at
the tested position. Possibility to choose among: "cr", "par", "anc" or
"biall". For details look at \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument if the user wants to allow QTLs with an
ancestral effect. \code{Interger matrix} representing the results of a parents
genotypes clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{cofactors}{Vector of \code{character} marker or in between marker
positions names. Default = NULL.}

\item{cof.Qeff}{\code{Character} vector indicating for each cofactor position
the type of QTL effect from: "cr", "par", "anc" and "biall".}

\item{chg.Qeff}{\code{Logical} value. If \code{chg.Qeff = TRUE}.
The type of QTL effect of the tested position will change for the one
specified in \code{cof.Qeff} when the function enter the region of a cofactor
delimited by \code{window}. Default = FALSE.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run function in parallel user must
provide cluster in the \code{cluster} argument. \strong{Parallelization is
only available for HRT (linear) models \code{VCOV = "h.err"}}.
Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}
}
\value{
Return:

\item{CIM }{\code{Data.frame} of class \code{QTLprof} with five columns :
1) QTL marker or in between position names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-values)}
}
\description{
Compute a QTL profile with cofactors position having different type of QTL
effects.
}
\details{
\strong{WARNING!} The computation of random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can sometimes fail. This could be
due to singularities due to a strong correlation between the QTL term(s) and 
the polygenic term. This situation can appear in the parental model.
the error can also sometimes come from the \code{asreml()} function. From
our experience, in that case, trying to re-run the function one or two times
allow to obtain a result.
}
\examples{

\dontrun{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)

mppData <- USNAM_mppData
mppData_bi <- USNAM_mppData_bi
par.clu <- USNAM_parClu

# Equalize the list of markers of the two mppData objects and par.clu

com.mk.list <- intersect(mppData$map$mk.names, mppData_bi$map$mk.names)

mppData <- mppData_subset(mppData = mppData, mk.list = com.mk.list)
mppData_bi <- mppData_subset(mppData = mppData_bi, mk.list = com.mk.list)
par.clu <- par.clu[rownames(par.clu) \%in\% com.mk.list, ]

SIM <- mpp_SIM(mppData = mppData)
cofactors <- QTL_select(SIM)[, 1]

CIM <- MQE_CIM(mppData = mppData, mppData_bi = mppData_bi, Q.eff = "cr",
               par.clu = par.clu, cofactors = cofactors,
               cof.Qeff = c("anc", "par", "biall"))

plot_QTLprof(CIM)

}
                               
}
\seealso{
\code{\link{mppData_form}}, \code{\link{mpp_SIM}},
\code{\link{parent_cluster}}, \code{\link{QTL_select}}
}
\author{
Vincent Garin
}
