% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQE_forward.R
\name{MQE_forward}
\alias{MQE_forward}
\title{Forward regression with different type of QTL effects}
\usage{
MQE_forward(mppData = NULL, mppData_bi = NULL, Q.eff, par.clu = NULL,
  VCOV = "h.err", threshold = 3, window = 20, parallel = FALSE,
  cluster = NULL, verbose = TRUE)
}
\arguments{
\item{mppData}{An IBD object of class \code{mppData}
See \code{\link{mppData_form}} for details. Default = NULL.}

\item{mppData_bi}{Required IBS object of class \code{mppData} if the user
wants to allow QTLs with a bi-allelic effect. \strong{The list of marker must
be strictly the same as the one of \code{mppData}.} Default = NULL.}

\item{Q.eff}{\code{Character} vector of possible QTL effects the user want to
test. Elements of Q.eff can be "cr", "par", "anc" or "biall". For details
look at \code{\link{mpp_SIM}}.}

\item{par.clu}{Required argument if the user wants to allow QTLs with an
ancestral effect. \code{Interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{threshold}{\code{Numeric} value representing the -log10(p-value) threshold
above which a position can be considered as significant. Default = 3.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run function in parallel user must
provide cluster in the \code{cluster} argument. \strong{Parallelization is
only available for HRT (linear) models \code{VCOV = "h.err"}}.
Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}

\item{verbose}{\code{Logical} value indicating if the steps of the
MQE_forward should be printed. It will not affect the printing of the other
functions called by \code{mpp_proc()}, especially the printing of
\code{asreml()}. Default = TRUE.}
}
\value{
\item{QTL.list }{\code{Data.frame} with six columns :
1) QTL marker or in between position names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; 5) -log10(p-values); and 6) type of
QTL incidence matrix of the selected positions}
}
\description{
Determines a multi-QTL effect (MQE) model using a forward regression.
}
\details{
The included QTL position can have different type of QTL effects at different
loci. At each step (new added position), the function compute QTL profiles
using one type of QTL effect specified in \code{Q.eff} for the tested
position. Let us assume that the user want to allow the QTL to have a
parental, ancestral or bi-allelic effect. Therefore, the function will
calculate three profiles. Then it will select in each profile the most
significant position if a position is above the \code{threshold} value.
The position and its QTL effect that increase the most the global adjusted R
squared (\code{MQE_R2}) will be selected as the new QTL position and included
in the list of cofactors. The function continues to iterate until no position
is significant anymore.

\strong{WARNING!(1)} The computation of \code{MQE_forward()} function using
mixed models (all models with \code{VCOV} different than \code{"h.err"})
is technically possible but can be irrealistic in practice due to a reduced
computer power. Since a mixed model is computed at each single position it
can take a lot of time. From our estimation it can take between 20 to 50
times more time than for the linear model (HRT). If the number of detected
QTL is supposed to be small (until 5) it could still be feasible.

\strong{WARNING!(2)} The computation of random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can sometimes fail. This could be
due to singularities due to a strong correlation between the QTL term(s) and 
the polygenic term. This situation can appear in the parental model.
the error can also sometimes come from the \code{asreml()} function. From
our experience, in that case, trying to re-run the function one or two times
allow to obtain a result.
}
\examples{

\dontrun{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)

mppData <- USNAM_mppData
mppData_bi <- USNAM_mppData_bi
par.clu <- USNAM_parClu


# equalize the list of markers of the two mppData objects and par.clu

com.mk.list <- intersect(mppData$map$mk.names, mppData_bi$map$mk.names)

mppData <- mppData_subset(mppData = mppData, mk.list = com.mk.list)
mppData_bi <- mppData_subset(mppData = mppData_bi, mk.list = com.mk.list)
par.clu <- par.clu[rownames(par.clu) \%in\% com.mk.list, ]


QTL <- MQE_forward(mppData = mppData, mppData_bi = mppData_bi,
                   Q.eff = c("par", "anc", "biall"), par.clu = par.clu)
                   
R2 <- MQE_R2(mppData = mppData, QTL = QTL[, 1], Q.eff = QTL[, 5],
par.clu = par.clu)

}

}
\seealso{
\code{\link{mppData_form}}, \code{\link{mpp_perm}},
\code{\link{mpp_SIM}},
\code{\link{MQE_R2}}, \code{\link{parent_cluster}},
\code{\link{USNAM_parClu}}
}
\author{
Vincent Garin
}
