% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_clu_chg.mppData.R
\name{par_clu_chg.mppData}
\alias{par_clu_chg.mppData}
\title{Change the parent clustering in a \code{mppData} object}
\usage{
par_clu_chg.mppData(mppData, par.clu)
}
\arguments{
\item{mppData}{An object of class \code{mppData}. The \code{mppData} object
must have been processed with \code{\link{parent_cluster.mppData}}}

\item{par.clu}{\code{Interger matrix} representing the results of a
parents genotypes clustering. The columns represent the parental lines and
the rows the markers. The columns names must be the same as the parents
list of the mppData object. The rownames must be the same as the map marker
list of the mppData object. At a particular position, parents with the same
value are assumed to inherit from the same ancestor. for more details,
see \code{\link{par_clu}}.}
}
\value{
\code{mppData} object with the new parent clustering information.
}
\description{
Change the parent clustering in a \code{mppData} object.
}
\examples{

data(mppData)
data(par_clu)

mppData <- par_clu_chg.mppData(mppData = mppData, par.clu = par_clu)


}
\seealso{
\code{\link{parent_cluster.mppData}}, \code{\link{par_clu}}
}
\author{
Vincent Garin
}
