% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_cluster.R
\name{parent_cluster}
\alias{parent_cluster}
\title{Clustering of parental lines}
\usage{
parent_cluster(haplo.map, consensus.map, marker.data, na.strings = NA,
  w1 = "kernel.exp", w2 = "kernel.unif", step.size = 10000, window,
  K = 10, simulation.type = "equi", simulation.Ng = 50,
  simulation.Nrep = 3, threshold.quantile = 95, plot = TRUE,
  plot.loc = getwd())
}
\arguments{
\item{haplo.map}{Three columns \code{data.frame} with: 1)
marker identifiers; 2) chromosomes; and 3) \code{numeric} position in
centi-Morgan. This argument represents the genetic map of the parents markers
used for the clustering procedure. It can contain as many markers as possible
to increased the available information for clustering.}

\item{consensus.map}{\code{Data.frame} with the same format as
\code{haplo.map}. This is the map used for the QTL analysis.}

\item{marker.data}{\code{Character} marker \code{matrix} representing the
parent scores for the markers present in \code{haplo.map}. \strong{the
rownames of \code{marker.data} and the marker list of \code{haplo.map} must
be strictly equivalents.}}

\item{na.strings}{\code{Character} expression for marker scores missing value
(".", "NA", "-"). Default = NA.}

\item{w1}{The w1 weight function in the Li&Jyang similarity score.
Possible values are "kernel.const", "kernel.exp", "kernel.gauss",
"kernel.unif", "kernel.laplace" or "kernel.null". Default = "kernel.exp".}

\item{w2}{The w2 weight function in the Li&Jyang similarity score.
Possible values are "kernel.const", "kernel.exp", "kernel.gauss",
"kernel.unif", "kernel.laplace" or "kernel.null". default = "kernel.unif".}

\item{step.size}{\code{Numeric} value. If the length between two marker
position of consensus.map is bigger than step.size then clusthaplo will
caculate a clustering results of the parental genotype at step.size
interval(s). The default value is set to 10000. This means that no ancestral
class will be inferred in between position of the consensus map.}

\item{window}{\code{Numeric} value for the size of the window used for
clustering in centi-Morgan. The clustering procedure is done for the position
that is in the centre of the window taking marker scores within the window
into consideration.}

\item{K}{A positive integer representing the number of markers in a window
below which the kinship data will be used. Default = 10.}

\item{simulation.type}{The type of simulation used for the training.
One of "equi" or "mosaic". Default = "equi".}

\item{simulation.Ng}{The number of intermediary generations to simulate
for the training (only relevant for "mosaic"). Default = 50.}

\item{simulation.Nrep}{The number of replicates to simulate for the training.
default value = 3.}

\item{threshold.quantile}{The quantile to use to select the threshold
automatically. It must be a plain integer xx with 80 <= xx <= 100.
Default = 95.}

\item{plot}{\code{Logical} value indicating if the plot of the clustering
results must be saved at the location specified in argument \code{plot.loc}.
Default = TRUE.}

\item{plot.loc}{Path where a folder will be created to save the plot of
the clustering results. By default the function uses the current working
directory.}
}
\value{
Return:

\code{List} with the following objects

\item{par.clu }{\code{Integer matrix} with rows repersenting marker or
in between positions and column corresponding to the parents. At a single
marker position, parents with the same value were clustered in the same
ancestral group.}

\item{av.cl }{Average number of cluster. }

Plot of the ancestral haplotypes are saved in a folder at the location
specified in \code{plot.loc} if \code{plot = TRUE}.
}
\description{
Local clustering of the parental lines obtained using the R package
clushaplo (Leroux et al. 2014).
}
\details{
This function is a wrapper of \code{clusthaplo} package functions.
It can only be run after package Clusthaplo has been installed.
It can be found there:
\url{https://cran.r-project.org/src/contrib/Archive/clusthaplo/}. A
visualisation of ancestral haplotype blocks can be obtained setting
\code{plot = TRUE}. The plots will be saved at the location specified
in \code{plot.loc}. For the moment, due to the difficulty to have the
package \code{mppR}, \code{clusthaplo} and \code{RHmm} functioning
simultaneously, the clutering based on hidden Markov model is unavailable.
Therefore the default method for clustering is threshold (Leroux et al. 2014).
}
\examples{

\dontrun{

library(clusthaplo)

data(USNAM_geno)
data(USNAM_mppData)
data(USNAM_map)

# Parents genotypes and map
geno.par <- USNAM_geno[1:6, ]
geno.par <- t(geno.par)
map.par <- USNAM_map

# Map obtained after mpp.data object formation and potential
# inference of position in between existing markers
map <- USNAM_mppData$map[, -3]


par.clu <- parent_cluster(haplo.map = map.par, consensus.map = map,
marker.data = geno.par, step.size = 1000, window = 25, plot = TRUE,
plot.loc = getwd())


par.clu$av.cl # Average number of inferred ancestral cluster along the genome

# Collect the parent clustering results for the analysis

par.clu <- par.clu$par.clu 

SIM <- mpp_SIM(mppData = USNAM_mppData, Q.eff = "anc", par.clu = par.clu)
plot_QTLprof(SIM)

}


}
\references{
Leroux, D., Rahmani, A., Jasson, S., Ventelon, M., Louis, F., Moreau, L.,
& Mangin, B. (2014). Clusthaplo: a plug-in for MCQTL to enhance QTL detection
using ancestral alleles in multi-cross design. Theoretical and Applied
Genetics, 127(4), 921-933.
}
\seealso{
\code{\link{parent_clusterCheck}}
}
\author{
Vincent Garin
}
