% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_update.mppData.R
\name{pedigree_update.mppData}
\alias{pedigree_update.mppData}
\title{Modify pedigree information in a mppData object}
\usage{
pedigree_update.mppData(mppData, pedigree)
}
\arguments{
\item{mppData}{An object of class \code{mppData}. The \code{mppData} object
must have been processed with \code{\link{QC.mppData}}}

\item{pedigree}{Four columns \code{data.frame}: 1) the type of
genotype: "offspring" for the last genration and "founder" for genotype above
the offspring in the pedigree; 2) the genotype indicator of offsrping and
founders; 3-4) the parent 1 (2) of each line. \strong{The genotypes
identifiers of the offspring must be identical to \code{mppData$geno.id}.
The row giving the pedigree of an individual must appear before any row
where that individual appears as a parent.}}
}
\value{
Return:

\item{mppData}{New \code{mppData} object with completed pedigree information.}
}
\description{
Modify the pedigree information of a \code{mppData} object.
}
\details{
In the formation of the \code{mppData} object, the functions only
uses the parental relationships specified in the \code{par.per.cross}
argument. However, if pedigree relationship among parents are know by the
user a more complete pedigree information can be provided to the
\code{mppData} object. This pedigree information can be given via the
\code{pedigree} argument of the \code{pedigree_update.mppData()} function.
}
\examples{

data(mppData)
pedigree <- mppData$ped.mat

# Let us assume for example that parent CML103 and CML322 share a
# common parent (A1). Then we can write a new pedigree adding this
# information.

founder.info <- data.frame(rep("founder", 2), c("CML103", "CML322"),
                           rep("A1", 2), c("A2", "A3"))
colnames(founder.info) <- colnames(pedigree)

pedigree.new <- rbind(founder.info, pedigree)

mppData <- pedigree_update.mppData(mppData = USNAM_mppData,
pedigree = pedigree.new)


}
\seealso{
\code{\link{QC.mppData}}, \code{\link{subset.mppData}}
}
\author{
Vincent Garin
}
