% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genEffects.R
\name{plot_genEffects}
\alias{plot_genEffects}
\title{plot of QTL genetic effects}
\usage{
plot_genEffects(Qprof, QTL = NULL, main = "QTL genetic effects plot")
}
\arguments{
\item{Qprof}{Object of class \code{QTLprof} returned by the function
\code{\link{mpp_SIM}} or \code{\link{mpp_CIM}} with argument
\code{est.gen.eff = TRUE}.}

\item{QTL}{Optional argument. Object of class \code{QTLlist} representing a
list of selected position obtained with the function \code{\link{QTL_select}}
or vector of \code{character} marker or inbetween marker positions names.
These positions will be plotted on the graph. Default = NULL.}

\item{main}{Title of the graph. Default =  "QTL genetic effects plot".}
}
\description{
Visualisation of the genome-wide significance of the QTL effect per cross or
per parents. It uses the results of a SIM or CIM profile from a
cross-specific, parental and ancestral model. candidate QTL positions passed
to the argument \code{QTL} are also drawn on the graph.
}
\details{
Plot the significance of the cross-specific or parental components of the QTL
effects. White colour means no significant effect. For a cross-specific
QTL profile: Red color means that the allele coming from parent A(1)
increases the phenotypic value and parent B(2) decreases it
and blue that parent A(1) decreases the trait and parent B(2) increases it.

For a parental or an ancestral QTL profile: Red colour means a significant
negative difference with respect to the first parent (ancestral group)
taken as reference, Blue represent a significant positive difference.
The colour intensity increase with the significance of the effect.
There are five colour intensities according to the p-value of the QTL
effect: 0.05<p-val<0.01; 0.01<p-val<0.001; 0.001<p-val<0.0001;
0.0001<p-val<0.00001 and p-val< 0.00001.

The dashed lines indicate the position that have been introduce in the
argument \code{QTL}.
}
\examples{

data(USNAM_mppData)

SIM <- mpp_SIM(mppData = USNAM_mppData, Q.eff = "cr", est.gen.eff = TRUE)
QTL <- QTL_select(SIM)

plot_genEffects(Qprof = SIM) # without QTL positions

plot_genEffects(Qprof = SIM, QTL = QTL) # with QTL positions

}
\author{
Vincent Garin
}
\seealso{
\code{\link{mpp_CIM}}, \code{\link{mpp_SIM}}, \code{\link{QTL_select}}
}

