% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{print.summary.mppData}
\alias{print.summary.mppData}
\title{Print summary.mppData summary.QeffRes or summary.QR2Res object}
\usage{
\method{print}{summary.mppData}(x, ...)
}
\arguments{
\item{x}{object of class \code{summary.mppData}, \code{summary.QeffRes} or,
\code{summary.QR2Res}}

\item{...}{Ignored.}
}
\description{
Print summary.mppData summary.QeffRes or summary.QR2Res object
}
\examples{

data(mppData)
sum.mppData <- summary(mppData)
print(sum.mppData)

data(mppData)
SIM <- mpp_SIM(mppData)
QTL <- QTL_select(SIM)
QTL.effects <- QTL_gen_effects(mppData = mppData, QTL = QTL, Q.eff = "cr")
sum.QeffRes <- summary(QTL.effects)
print(sum.QeffRes)

data(mppData)
SIM <- mpp_SIM(mppData)
QTL <- QTL_select(SIM)
Q_R2 <- QTL_R2(mppData = mppData, QTL = QTL, Q.eff = "cr")
summary(Q_R2)

}
