% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.QeffRes.R
\name{print.summary.QeffRes}
\alias{print.summary.QeffRes}
\title{Print summary.QeffRes object}
\usage{
\method{print}{summary.QeffRes}(x, ...)
}
\arguments{
\item{x}{object of class \code{summary.QeffRes}}

\item{...}{Ignored.}
}
\description{
Print summary.QeffRes object
}
\examples{

data(mppData)
SIM <- mpp_SIM(mppData)
QTL <- QTL_select(SIM)
QTL.effects <- QTL_gen_effects(mppData = mppData, QTL = QTL, Q.eff = "cr")
sum.QeffRes <- summary(QTL.effects)
print(sum.QeffRes)

}
