% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.QR2Res.R
\name{print.summary.QR2Res}
\alias{print.summary.QR2Res}
\title{Print summary.QR2Res object}
\usage{
\method{print}{summary.QR2Res}(x, ...)
}
\arguments{
\item{x}{object of class \code{summary.QR2Res}}

\item{...}{Ignored.}
}
\description{
Print summary.QR2Res object
}
\examples{

data(mppData)
SIM <- mpp_SIM(mppData)
QTL <- QTL_select(SIM)
Q_R2 <- QTL_R2(mppData = mppData, QTL = QTL, Q.eff = "cr")
sum.QR2Res <- summary(Q_R2)
print(sum.QR2Res)

}
