% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_GenotypingError.R
\name{QC_GenotypingError}
\alias{QC_GenotypingError}
\title{Detect genotyping errors}
\usage{
QC_GenotypingError(mk.mat, parallel = FALSE, cluster = NULL)
}
\arguments{
\item{mk.mat}{\code{Character} marker score \code{matrix} with genotype as
row and marker as column. \strong{Rows and columns names must be the genotype
and marker identifiers respectively. Marker scores must be coded using one
letter per allele. For example, AA, CC, GG, TT, AC, AG, AT, CA, CG, CT,
GA, GC, GT, TA, TC, TG. Missing values must be coded \code{NA}.}}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}
}
\value{
Return:

\item{prob.mk}{\code{Character} vector containing the name of marker with
more than two alleles.}
}
\description{
Checks if markers present more than two alleles.
}
\examples{

data(USNAM_geno)

mk.mat <- USNAM_geno
mk.mat[1, 10] <- "TC" 
QC_GenotypingError(mk.mat)

}
\seealso{
\code{\link{QC_MAF}}, \code{\link{QC_missing}}
}
\author{
Vincent Garin
}
