% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_hetero.R
\name{QC_hetero}
\alias{QC_hetero}
\title{Marker heterozigosity rate}
\usage{
QC_hetero(mk.mat, parallel = FALSE, cluster = NULL)
}
\arguments{
\item{mk.mat}{\code{Character} marker score \code{matrix} with genotypes as
row and markers as column. \strong{Rows and columns names must be the genotype
and marker identifiers respectively. Marker scores must be coded using one
letter per allele. For example, AA, CC, GG, TT, AC, AG, AT, CA, CG, CT,
GA, GC, GT, TA, TC, TG. Missing values must be coded \code{NA}.}}

\item{parallel}{Logical value specifying if the function should be executed
in parallel on multiple cores. Default value = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}
}
\value{
Return:

\item{het.rate}{\code{Numeric} vector specifying the heterozygosity rate
for each marker.}
}
\description{
Calculates marker heterozigosity rate.
}
\examples{

data('USNAM_geno')

hetero <- QC_hetero(mk.mat = USNAM_geno[7:dim(USNAM_geno)[1], ])

}
\author{
Vincent Garin
}

