% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_MAF.R
\name{QC_MAF}
\alias{QC_MAF}
\title{Minor allele frequency}
\usage{
QC_MAF(mk.mat, cross.ind = NULL, parallel = FALSE, cluster = NULL)
}
\arguments{
\item{mk.mat}{\code{Character} marker score \code{matrix} with genotypes as
row and markers as column. \strong{Rows and columns names must be the genotype
and marker identifiers respectively. Marker scores must be coded using one
letter per allele. For example, AA, CC, GG, TT, AC, AG, AT, CA, CG, CT,
GA, GC, GT, TA, TC, TG. Missing values must be coded \code{NA}.}}

\item{cross.ind}{\code{Character} vector indicating to which cross each
genotype belong. If \code{cross.ind = NULL}, the function do not consider cross
subdivisions and only calculate MAF at the population level. Default = NULL.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}
}
\value{
Return:

Object of class mafRes MAF containing the following element(s).
If cross.ind = NULL

\item{MAF.pop}{Vector of marker allele MAF at the
population level. \code{NA} if all markers scores are missing.}

If cross.ind is not NULL

\item{MAF.pop}{Vector of marker allele MAF at the
population level. \code{NA} if all markers scores are missing.}

\item{MAF.cr}{\code{Matrix} of marker allele MAF within crosses.
\code{NA} if all markers scores are missing.}
}
\description{
Computes minor allele frequency (MAF) at the whole population and
within crosses.
}
\examples{

data(USNAM_geno)
cross.ind <- substr(rownames(USNAM_geno)[7:dim(USNAM_geno)[1]], 1, 4)

MAF <- QC_MAF(mk.mat = USNAM_geno[7:dim(USNAM_geno)[1], ],
             cross.ind = cross.ind)

}
\author{
Vincent Garin
}
\seealso{
\code{\link{QC_tagMAFCr}}
}

