% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_matchMarker.R
\name{QC_matchMarker}
\alias{QC_matchMarker}
\title{Match markers in the genotype matrix and the map}
\usage{
QC_matchMarker(mk.mat, map)
}
\arguments{
\item{mk.mat}{Marker score \code{matrix} or \code{data.frame} with genotypes
as row and markers as column. \strong{Marker names must be column names of the
matrix.}}

\item{map}{\code{Data.frame} \strong{with first column containing the marker
names}.}
}
\value{
Return:

\code{List} containing the following objects:

\item{new.mk.mat}{Marker matrix with only the common list of
marker between the marker matrix and the map in the same order as in the map.}

\item{new.map}{Map with only the common markers list.}

\item{mk.mat.only}{Markers present in the marker matrix but not in the map.}

\item{map.only}{Markers present in the map but not in the marker matrix.}
}
\description{
Determines which marker are common between the genotypes marker matrix and
makers present in the map. Return the marker matrix and the map that
correspond to these markers.
}
\examples{

data(USNAM_map)
data(USNAM_geno)

# remove one marker from the map
map.red <- USNAM_map[-1,]

match <- QC_matchMarker(mk.mat = USNAM_geno, map = map.red)

# Marker only in the map
match$map.only

# Marker only in the marker matrix
match$mk.mat.only

# Get the new map and marker matrix
new.mk.mat <- match$new.mk.mat
new.map <- match$new.map

rm(match)

}
\seealso{
\code{\link{QC_matchGeno}}
}
\author{
Vincent Garin
}
