% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_minCrSize.R
\name{QC_minCrSize}
\alias{QC_minCrSize}
\title{Minimum cross size}
\usage{
QC_minCrSize(mk.mat, cross.ind, n.lim = 15)
}
\arguments{
\item{mk.mat}{Marker score \code{matrix} or \code{data.frame} with genotype as
row and markers as column.}

\item{cross.ind}{\code{Character} vector indicating to which cross each
genotype belong.}

\item{n.lim}{\code{Numeric} value specifying the minimum cross size.
Default = 15.}
}
\value{
Return:

\item{mk.mat.red}{Reduced marker matrix with too small crosses removed.}
}
\description{
Removes the cross that do not have the minimum critical size.
}
\examples{

data('USNAM_geno')

USNAM_geno.red <- USNAM_geno[-c(7:101), ] # remove 95 genotypes in the first cross

cross.ind <- substr(rownames(USNAM_geno.red), 1, 4)

mk.mat.red <- QC_minCrSize(mk.mat = USNAM_geno.red, cross.ind = cross.ind)

}
\author{
Vincent Garin
}
