% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_missing.R
\name{QC_missing}
\alias{QC_missing}
\title{Missing value for marker or genotype}
\usage{
QC_missing(mk.mat, threshold = 0.1, MARGIN = 2)
}
\arguments{
\item{mk.mat}{\code{Character} marker score \code{matrix} with genotype as
row and marker as column. \strong{Rows and columns names must be the genotype
and marker identifiers respectively. Missing values must be coded \code{NA}.}}

\item{threshold}{\code{Numeric} value representing the missing values rate
above which a marker or a genotype is considered as problematic.
Default = 0.1.}

\item{MARGIN}{\code{Numeric} value giving the subscript which the function
will be applied over. If MARGIN = 1,  the function looks at the genotypes
(rows). If MARGIN = 2, the function looks at the markers (columns).
Default = 2.}
}
\value{
Return:

\item{mis.ind }{Three columns \code{data.frame} with: 1) marker or
genotype names; 2) column or row positions in the matrix; and
3) The percentages of missing values.}
}
\description{
Identify markers or genotypes with a missing rate higher than the
specified \code{threshold}.
}
\examples{

data('USNAM_geno')

# missingness of markers
QC_missing(mk.mat = USNAM_geno)

# missingness of genotypes
QC_missing(mk.mat = USNAM_geno, threshold = 0.2, MARGIN = 1)

}
\author{
Vincent Garin
}
