% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_effect_plot.R
\name{QTL_effect_plot}
\alias{QTL_effect_plot}
\title{Plot QTL effect accross model}
\usage{
QTL_effect_plot(mppData, trait = 1, QTL, sum_zero = TRUE, ref.par = NULL,
  trait.lab = "trait", text.size = 16, output.loc = getwd(),
  plot_label = "Qeff_tab")
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the \code{mppData} object should be used. Default = 1.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected position obtained with the function \code{\link{QTL_select}} or
vector of \code{character} marker or in between marker positions names.}

\item{sum_zero}{Optional \code{Logical} value specifying if the QTL effect of
a parental or an ancestral model should be caculated using the sum to zero
constraint. Default = TRUE.}

\item{ref.par}{Optional \code{Character} expression defining the parental
allele that will be used as reference for the parental model. For the
ancestral model, the ancestral class containing the reference parent will be
set as reference. \strong{This option can only be used if the MPP design is
composed of a unique connected part}. Default = NULL.}

\item{trait.lab}{\code{Character} expression indicating the name of the trait.
Default = 'trait'.}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 16.}

\item{output.loc}{Path to a folder where the results will be saved.
By default the function uses the current working directory.}

\item{plot_label}{\code{Character} for the mane of the plot. plot_label.pdf
will be saved at output.loc. Default = 'Qeff_tab'.}
}
\value{
Return:

\item{Qeff_tab}{\code{List} with one element per QTL with the QTL effects
and p-values for the different models that where plotted.}

A file plot_label.pdf saved at output.loc containing the plot of the QTL
effects.
}
\description{
Function to visualise and compare the effects of QTLs accross the different
possible models for the QTL effect (cross-specific, parental, ancestral,
bi-allelic).
}
\details{
The function calculates the effects of the different QTL positions in the
four possible models (cross-specific, parental, ancsetral, bi-allelic). Then
the effects are ploted using buble. The size of the buble is proportional to
the significance of the effect. The intensity of the colour is proportional
to the size of the effect. The red colour means that the allele has a negative
effect on the trait and the blue that it has a positive effect.

The effects are presented using the crosses as reference. For each cross,
the parent A and B are given. The effect of the cross-specific model are
represented per parent. The two parents have the same contribution because
the effect are estimated within cross contrasting the two parent effects.
One parent has a positive contribution, the other the same negative
contribution.

Then the effect of the parental model are plotted. By default (sum_zero = TRUE)
these effects are estimated using a sum to zero constraint. This means that within a
connected part the sum of the parental allelic effect is zero. These effects
are plotted using the parent per cross as reference. A colour index indicate
which parent are the shared between crosses.

The ancestral effect are also estimated using a sum to zero constraint by
default. The ancestral effect are translated at the parental level. Parents
belonging to the same ancestral group get the same effect.

Finally, the bi-allelic effect are also represented. The bi-allelic effects
are also estimated using a sum to zero constraint.

An alternative is to estimate the QTL effect of the parental, ancestral,
and bi-allelic model using a reference allele set to 0 (sum_zero = FALSE).
In that case, the most frequent parental, ancestral and SNP allele will
be used as reference. For the parental and the ancestral model it is also
possible to fix the reference allele using argument ref.par.
}
\examples{

\dontrun{

data(mppData)

SIM <- mpp_SIM(mppData = mppData)
QTL <- QTL_select(SIM)

# Specify a location where your results will be saved
my.loc <- "C:/.../..."

MQE <- QTL_effect_plot(mppData = mppData, QTL = QTL, output.loc = my.loc)
                 
}

}
\author{
Vincent Garin
}
