% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_genEffects.R
\name{QTL_genEffects}
\alias{QTL_genEffects}
\title{QTL genetic effects}
\usage{
QTL_genEffects(mppData, QTL = NULL, Q.eff = "cr", par.clu = NULL,
  VCOV = "h.err", const = "set.0", par.ref = NULL)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected position obtained with the function \code{\link{QTL_select}} or
vector of \code{character} marker or inbetween marker positions names.
Default = NULL.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. For more details see \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument for the ancesral model \code{(Q.eff = "anc")}.
\code{interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{const}{\code{Character} expression indicating the type of contstraint
used for the estimation of the genetic effect of the HRT (linear) model
(\code{VCOV = "h.err"}) for the parental or ancestral models
(\code{Q.eff = "par" or "anc"}). If const = "set.0", the value of the
parent or the ancestral class containing the parent specified in argument
(\code{par.ref}) is set to 0. If const = "sum.0", the sum of
the parental (ancestral) effects is constrained to sum to 0.
Default = "set.0".}

\item{par.ref}{\code{Character} expression indicating the parent or the
ancestral class containing the specified parent that will be set as reference
for the computation of the genetic effect. By default
the function will use the first parent of the \code{mppData$parents} list.}
}
\value{
Return:

\item{results}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects per cross or parent.}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}
\item{For cross-specific model, parent with the positive additive effects.}
\item{Allele scores of the parents if \code{geno.par} is non NULL
in the \code{mppData} object.}

}

}
}
\description{
Computes a multi-QTL model with a list QTL candidates(\code{QTL}) and return
the decomposed QTL effects per cross or per parents.
}
\details{
This function computes for each QTL position the genetic effects of the
cross, parental, ancestral or SNP allele components. For the cross-specific
model (\code{Q.eff = "cr"}), the genetics effects represent the substitution
effect of an single allele from the parent 2 (or B) with respect to an allele
coming from the parent 1 or A. All effects are given in absolute value with
the parent that cary the positive allele.

For the parental or ancestral models (\code{Q.eff = "par"
or "anc"}), the genetic effects represent the effect of a single parental
(ancestral) allele with respect to a reference allelele. The reference
parental (ancestral) allele can be specified in the argument
\code{par.ref}. For the ancestral model, the ancestral class inferred
for the specified parent will be used as reference. For the parental and
ancestral model computed with the homogeneous residual term assumption
(\code{VCOV = "h.err"}), it is also possible to estimate one effect for
each parental (ancestral) allele with the constraint that these effects
sum to zero. This can be done using (\code{const = "sum.0"}).

For the bi-allelic model (\code{Q.eff = "biall"}), the genetic effects
represent the effects of a single allele copy of the most frequent allele.

\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.
}
\examples{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)
par.clu <- USNAM_parClu

# QTL candidates

SIM <- mpp_SIM(USNAM_mppData)
QTL <- QTL_select(SIM)

# cross-specific model

QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "cr")
QTL.effects

# parental model

# constraint(set one parents as reference - CML322)
QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "par",
                              const = "set.0", par.ref = "CML322")
QTL.effects

# constraint(sum to zero)
QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "par",
                              const = "sum.0")
QTL.effects

# ancestral model

# constraint(set one ancestral class as reference - M37W)
QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "anc",
                              par.clu = par.clu, const = "set.0", par.ref = "M37W")
QTL.effects

# constraint(sum to zero)
QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "anc",
                              par.clu = par.clu, const = "sum.0")
QTL.effects

# bi-allelic model

SIM <- mpp_SIM(USNAM_mppData_bi, Q.eff = "biall")
QTL <- QTL_select(SIM)


QTL.effects <- QTL_genEffects(mppData = USNAM_mppData_bi, QTL = QTL,
Q.eff = "biall")

QTL.effects

}
\author{
Vincent Garin
}
\seealso{
\code{\link{mppData_form}}, \code{\link{parent_cluster}},
\code{\link{QTL_select}}, \code{\link{USNAM_parClu}}
}

