% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_pred_R2.R
\name{QTL_pred_R2}
\alias{QTL_pred_R2}
\title{Predicted QTL global and partial R squared}
\usage{
QTL_pred_R2(mppData.ts, mppData.vs, Q.eff = "cr", par.clu = NULL,
  VCOV = "h.err", QTL = NULL, within.cross = TRUE)
}
\arguments{
\item{mppData.ts}{An object of class \code{mppData} for the training set. See
\code{\link{mppData_form}} for details.}

\item{mppData.vs}{An object of class \code{mppData} for the validation set.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. For more details see \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument for the ancesral model \code{(Q.eff = "anc")}.
\code{interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected position obtained with the function \code{\link{QTL_select}} or
vector of \code{character} marker or in between marker positions names.
Default = NULL.}

\item{within.cross}{\code{Logical} value indicating if the predicted
R squared must be computed at the whole population level or within cross.
In the later case, predicted R squared are computed within cross and the
average is returned. Default = TRUE.}
}
\value{
Return:

\code{List} containing the following objects:

\item{glb.R2 }{ Global predicted R squared of all QTL terms.}

\item{glb.adj.R2 }{ Global predicted adjusted R squared of all QTL terms.}

\item{part.R2.diff }{ Vector of predicted partial R squared doing
the difference between the full model and a model minus the ith QTL.}

\item{part.adj.R2.diff }{ Vector of predicted partial adjusted R squared
doing the difference between the full model and a model minus the ith QTL.}

\item{part.R2.sg }{ Vector of predicted partial R squared using only the
ith QTL.}

\item{part.adj.R2.sg }{ Vector of predicted partial adjusted R squared using
only the ith QTL.}
}
\description{
Computes predicted R squared in a validation set using QTLs detected in a
training set.
}
\details{
Compute QTLs predicted R squared in a validation set (\code{mppData.vs}).
These QTLs have been previously detected in a training set
(\code{mppData.ts}). The global R squared is obtained using the Pearson
squared correlation between the observed trait values in the validation set
(y.vs) and predicted values using estimated QTL effects in the training set
(y.pred.vs = X.vs * B.ts). The correlation can be calculated within cross and
then averaged (\code{within.cross = TRUE}) or at the whole population level.

Partial R squared statistics are also calculated for each individual position.
Two types or partial R squared are computed. The first one making the
difference between the global R squared and the R squared computed without
the ith position (difference R squared). The second method only uses
the ith QTL for trait values predition (single R squared). All predicted R
squared can be adjusted using formula 2 from Utz et al. (2000) (for details
see documentation of the function \code{\link{QTL_R2}}).

\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.
}
\examples{

data(USNAM_mppData)

folds <- CV_partition(cross.ind = USNAM_mppData$cross.ind, k = 5)

mppData.ts <- mppData_subset(mppData = USNAM_mppData,
                             gen.list = folds[[1]]$train.set)

mppData.vs <- mppData_subset(mppData = USNAM_mppData,
                             gen.list = folds[[1]]$val.set)

SIM <- mpp_SIM(mppData = USNAM_mppData)
QTL <- QTL_select(SIM)

QTL_pred_R2(mppData.ts = mppData.ts, mppData.vs = mppData.vs, QTL = QTL)

}
\author{
Vincent Garin
}
\references{
Utz, H. F., Melchinger, A. E., & Schon, C. C. (2000). Bias and sampling error
of the estimated proportion of genotypic variance explained by quantitative
trait loci determined from experimental data in maize using cross validation
and validation with independent samples. Genetics, 154(4), 1839-1849.
}
\seealso{
\code{\link{mppData_form}}, \code{\link{parent_cluster}},
\code{\link{QTL_R2}}, \code{\link{QTL_select}}, \code{\link{USNAM_parClu}}
}

