% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_res_list.R
\name{QTL_res_list}
\alias{QTL_res_list}
\title{List of QTL results}
\usage{
QTL_res_list(mppData, MPP_out, trait, Q.eff, VCOV, res_file = c())
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{MPP_out}{Output from \code{\link{mpp_proc}}.}

\item{trait}{\code{character} indicator to specify the  trait name.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 'cr', 'par', 'anc', 'biall' or 'MQE'.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used.}

\item{res_file}{\code{data.frame} to store the QTL effects results. Default,
empty file.}
}
\value{
The results of \code{MPP_out} are appended to \code{res_file}.
}
\description{
Form a list of QTL results appending QTL effects results obtained during
different QTL detection procedure. The results are appended to
\code{res_file}.
}
\examples{

# not yet

}
\author{
Vincent Garin
}
