% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mppData.R
\name{summary.mppData}
\alias{summary.mppData}
\title{Print summary of mppData object}
\usage{
\method{summary}{mppData}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 \code{summary} method for object of class \code{mppData}. The summary of
the map information lists only the marker and not the added in between
positions.
}
\examples{

data(USNAM_mppData)
summary(USNAM_mppData)

}
\seealso{
\code{\link{mppData_form}}, \code{\link{USNAM_mppData}},
\code{\link{USNAM_mppData_bi}}
}
