% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.QR2Res.R
\name{summary.QR2Res}
\alias{summary.QR2Res}
\title{Summary of \code{QR2Res} object}
\usage{
\method{summary}{QR2Res}(object, ...)
}
\arguments{
\item{object}{An object of class \code{QR2Res} obtained with
function \code{\link{QTL_R2}}.}

\item{...}{Ignored.}
}
\description{
\code{summary} for object of class \code{QR2Res}.
}
\examples{

data(mppData)
SIM <- mpp_SIM(mppData)
QTL <- QTL_select(SIM)
Q_R2 <- QTL_R2(mppData = mppData, QTL = QTL, Q.eff = "cr")
summary(Q_R2)

}
\seealso{
\code{\link{QTL_R2}}
}
