% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMapMaker.R
\name{toMapMaker}
\alias{toMapMaker}
\title{Map into MapMaker format}
\usage{
toMapMaker(map)
}
\arguments{
\item{map}{Three columns \code{data.frame} with: 1) \code{character}
marker identifiers; 2) \code{numeric} integer chromosome indicators
(1, 2, 3,...); and 3) \code{numeric} positions in centi-Morgan.}
}
\value{
Return:

\item{new.map}{Map with the MapMaker format: \code{list} of matrix with one
element for each chromosome. Each single chromosome matrix is composed of four
columns: 1) marker identifiers; 2) dist.to: one lagged difference
positions in cM with Inf at first position; 3) dist.from: one lagged
difference positions in cM with Inf at last position ; 4) locus: marker
positions in cM.}
}
\description{
Transforms a map file with three columns (marker identifiers, chromosomes and
positions in cM) into MapMaker format.
}
\examples{

data(USNAM_map)

map_MapMaker <- toMapMaker(USNAM_map)

}
\author{
Vincent Garin
}

