\name{USNAM_map}
\alias{USNAM_map}
\docType{data}
\title{
  Reduced map maize US-NAM population
}
\description{
  Reduced map of the maize US nested association mapping (NAM) population
  (McMullen et al., 2009).
}
\usage{data(USNAM_map)}
\format{
\code{data.frame}
}
\details{
  Selection of 102 markers from the two first chromosomes of the Maize US-NAM
  population (McCullen et al., 2009). The data are available on www.panzea.org.
}
\source{
  \url{https://www.panzea.org}
}
\references{
  McMullen, M. D., Kresovich, S., Villeda, H. S., Bradbury, P., Li, H., Sun,
  Q., ... & Buckler, E. S. (2009). Genetic properties of the maize nested
  association mapping population. Science, 325(5941), 737-740.
  
}
\examples{
  data(USNAM_map)
}

\seealso{
\code{\link{USNAM_geno}}, \code{\link{USNAM_pheno}}
}

\keyword{datasets}