\name{USNAM_mppData}
\alias{USNAM_mppData}
\docType{data}
\title{
  IBD mppData object for MPP QTL analysis
}
\description{
  Identical by descent (IBD) mppData object of a sample of the Maize US nested
  association mapping (NAM) population for cross-specific, parental or
  ancestral models.
}
\usage{data(USNAM_mppData)}
\format{
  \code{mppData}
}
\details{
  see examples of the \code{\link{mppData_form}}.
}

\examples{
  data(USNAM_mppData)
}

\seealso{
\code{\link{mppData_form}}
}

\keyword{datasets}