\name{USNAM_mppData_bi}
\alias{USNAM_mppData_bi}
\docType{data}
\title{
  IBS mppData object for MPP QTL analysis
}
\description{
  Identical by state (IBS) mppData object of a sample of the maize US nested
  association mapping (NAM) population for bi-allelic models.
}
\usage{data(USNAM_mppData_bi)}
\format{
  \code{mppData}
}
\details{
  see examples of the \code{\link{mppData_form}}.
}

\examples{
  data(USNAM_mppData_bi)
}

\seealso{
\code{\link{mppData_form}}
}

\keyword{datasets}