% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpp_CIM.R
\name{mpp_CIM}
\alias{mpp_CIM}
\title{MPP composite interval maping}
\usage{
mpp_CIM(mppData, Q.eff = "cr", par.clu = NULL, VCOV = "h.err",
  cofactors = NULL, window = 20, est.gen.eff = FALSE, parallel = FALSE,
  cluster = NULL)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. For more details see \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument for the ancesral model \code{(Q.eff = "anc")}.
\code{interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{cofactors}{Object of class \code{QTLlist} representing a list of
selected position obtained with the function \code{\link{QTL_select}} or
vector of \code{character} marker or inbetween marker positions names.
Default = NULL.}

\item{window}{\code{Numeric} distance on the left an right of a cofactor
position where it is not included in the model. Default = 20.}

\item{est.gen.eff}{\code{Logical} value. if \code{est.gen.eff = TRUE},
the function will save the decomposed genetic effects per cross/parent.
These results can be ploted with the function \code{\link{plot_genEffects}}
to visualize a genome-wide decomposition of the genetic effects.
\strong{This functionality is ony available for the cross-specific,
parental and ancestral models.}
Default value = FALSE.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run function in parallel user must
provide cluster in the \code{cluster} argument. \strong{Parallelization is
only available for HRT (linear) models \code{VCOV = "h.err"}}.
Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}
}
\value{
Return:

\item{CIM }{\code{Data.frame} of class \code{QTLprof}. with five columns :
1) QTL marker or in between position names; 2) chromosomes;
3) Interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-val). And if
\code{est.gen.eff = TRUE}, p-values of the cross or parental QTL effects.}
}
\description{
Computes QTL models using cofactors with different possible assumptions
concerning the number of alleles at the QTL position and the variance
covariance structure (VCOV) of the model. For more details about the
different models, see documentation of the function \code{\link{mpp_SIM}}.
The function returns a -log10(p-value) QTL profile.
}
\details{
\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.
}
\examples{

# data
data(USNAM_mppData)
data(USNAM_mppData_bi)

# Cross-specific effect model
#############################

SIM <- mpp_SIM(mppData = USNAM_mppData, Q.eff = "cr", VCOV = "h.err")

cofactors <- QTL_select(Qprof = SIM, threshold = 3, window = 20)

CIM <- mpp_CIM(mppData = USNAM_mppData, Q.eff = "cr", VCOV = "h.err",
cofactors = cofactors, window = 20, est.gen.eff = TRUE)

plot_QTLprof(Qprof = CIM)
plot_genEffects(mppData = USNAM_mppData, Qprof = CIM, Q.eff = "cr")

\dontrun{

# Using multiple core

library(parallel)
n.cores <- detectCores()
cluster <- makeCluster((n.cores-1))

CIM <- mpp_CIM(mppData = USNAM_mppData, Q.eff = "cr", VCOV = "h.err",
cofactors = cofactors, window = 20, est.gen.eff = TRUE, parallel = TRUE,
cluster = cluster)

stopCluster(cl = cluster)

}

# Bi-allelic model
##################

SIM <- mpp_SIM(mppData = USNAM_mppData_bi, Q.eff = "biall", VCOV = "h.err")

cofactors <- USNAM_mppData_bi$map[c(15, 63), 1]

CIM <- mpp_CIM(mppData = USNAM_mppData_bi, Q.eff = "biall", VCOV = "h.err",
cofactors = cofactors, window = 20)

plot_QTLprof(Qprof = CIM, type = "h")
                               
}
\author{
Vincent Garin
}
\seealso{
\code{\link{mppData_form}}, \code{\link{mpp_SIM}},
\code{\link{parent_cluster}}, \code{\link{QTL_select}},
\code{\link{USNAM_parClu}}
}

