% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppData_subset.R
\name{mppData_subset}
\alias{mppData_subset}
\title{Subset mppData object}
\usage{
mppData_subset(mppData, mk.list = NULL, gen.list = NULL)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{mk.list}{\code{Character} vector of marker or in betwee position names,
\code{numeric} vector of marker or inbetween positions indicator,
or \code{logical} vector. Default = NULL.}

\item{gen.list}{\code{Character} vector of genotypes names, \code{numeric}
vector of genotypes, or \code{logical} vector. Default = NULL.}
}
\value{
Return:

\item{mppData_red}{Subseted \code{mppData} object with selected markers
or selected genotypes.}
}
\description{
Subsets \code{mppData} objects by markers (\code{mk.list}) or by genotypes
(\code{gen.list}).
}
\examples{

### marker subset

data(USNAM_mppData)

# Random selection of markers
mk.list <-  sample(USNAM_mppData$map[, 1], 50)
mppData_sub <- mppData_subset(mppData = USNAM_mppData, mk.list = mk.list)

# Selection marker of chromosome 1
mk.list <-  (USNAM_mppData$map[, 2] == 1)
mppData_sub <- mppData_subset(mppData = USNAM_mppData, mk.list = mk.list)

### genotype subset

# Random selection of genotypes
gen.list <-  sample(USNAM_mppData$geno.id, 200)
mppData_sub <- mppData_subset(mppData = USNAM_mppData, gen.list = gen.list)

# Selection of genotype from cross 2 and 5
crosses <- unique(USNAM_mppData$cross.ind)
gen.list <-  USNAM_mppData$geno.id[USNAM_mppData$cross.ind \%in\% crosses[c(2, 5)]]
mppData_sub <- mppData_subset(mppData = USNAM_mppData, gen.list = gen.list)

}
\seealso{
\code{\link{mppData_form}}, \code{\link{mppData_chgPheno}}
}
\author{
Vincent Garin
}
