% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQE_CIM.R
\name{MQE_CIM}
\alias{MQE_CIM}
\title{Composite interval maping for MQE model}
\usage{
MQE_CIM(mppData = NULL, trait = 1, Q.eff = "cr", VCOV = "h.err",
  cofactors = NULL, cof.Qeff, chg.Qeff = FALSE, window = 20,
  n.cores = 1)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the \code{mppData} object should be used. Default = 1.}

\item{Q.eff}{\code{Character} expression indicating the type of QTL effect at
the tested position. Possibility to choose among: "cr", "par", "anc" or
"biall". For details look at \code{\link{mpp_SIM}}. Default = "cr".}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{cofactors}{Vector of \code{character} marker or in between marker
positions names. Default = NULL.}

\item{cof.Qeff}{\code{Character} vector indicating for each cofactor position
the type of QTL effect from: "cr", "par", "anc" and "biall".}

\item{chg.Qeff}{\code{Logical} value. If \code{chg.Qeff = TRUE}.
The type of QTL effect of the tested position will change for the one
specified in \code{cof.Qeff} when the function enter the region of a cofactor
delimited by \code{window}. Default = FALSE.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{n.cores}{\code{Numeric}. Specify here the number of cores you like to
use. Default = 1.}
}
\value{
Return:

\item{CIM }{\code{Data.frame} of class \code{QTLprof} with five columns :
1) QTL marker or in between position names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-values)}
}
\description{
Compute a QTL profile with cofactors position having different type of QTL
effects.
}
\details{
\strong{WARNING!} The computation of random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can sometimes fail. This could be
due to singularities due to a strong correlation between the QTL term(s) and 
the polygenic term. This situation can appear in the parental model.
the error can also sometimes come from the \code{asreml()} function. From
our experience, in that case, trying to re-run the function one or two times
allow to obtain a result.
}
\examples{

data(mppData)

SIM <- mpp_SIM(mppData = mppData)
cofactors <- QTL_select(SIM)[, 1]

CIM <- MQE_CIM(mppData = mppData, Q.eff = "cr", cofactors = cofactors,
               cof.Qeff = c("anc", "par", "biall"))

plot(x = CIM)
                               
}
\seealso{
\code{\link{mppData_form}}, \code{\link{mpp_SIM}},
\code{\link{parent_cluster}}, \code{\link{QTL_select}}
}
\author{
Vincent Garin
}
