% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQE_genEffects.R
\name{MQE_genEffects}
\alias{MQE_genEffects}
\title{QTL genetic effects multi-QTL effect model}
\usage{
MQE_genEffects(mppData = NULL, mppData_bi = NULL, QTL = NULL, Q.eff,
  par.clu = NULL, VCOV = "h.err", ref.all.most = TRUE)
}
\arguments{
\item{mppData}{An IBD object of class \code{mppData}
See \code{\link{mppData_form}} for details. Default = NULL.}

\item{mppData_bi}{Required IBS object of class \code{mppData} if the user
wants to allow QTLs with a bi-allelic effect. \strong{The list of marker must
be strictly the same as the one of \code{mppData}.} Default = NULL.}

\item{QTL}{Vector of \code{character} markers or inbetween marker positions
names. Default = NULL.}

\item{Q.eff}{\code{Character} vector indicating for each QTL position the
type of QTL effect among: "cr", "par", "anc" and "biall". For details look at
\code{\link{mpp_SIM}}.}

\item{par.clu}{Required argument if the user wants to allow QTLs with an
ancestral effect. \code{interger matrix} representing the results of a parents
genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{ref.all.most}{\code{Logical} value specifying which parental
(ancestral) allele should be used as reference. If
\code{ref.all.most = TRUE}, within each connected part of the design,
the most used allele will be used as reference. Allele usage is first defined
in term of number of crosses where the allele segregates. Then, if two
alleles segregate in an equal number of crosses, we look at the total cross
sizes. If \code{ref.all.most = FALSE}, the less used allele will be used as
reference within each connected part. Default = TRUE.}
}
\value{
Return:

\item{results}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects per cross or parent.}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}
\item{For cross-specific model, parent with the positive additive effects.}
\item{For parental and ancestral model, indicator of connected part of the
design and reference.}
\item{Allele scores of the parents if \code{geno.par} is non NULL
in the \code{mppData} object.}

  }

}
}
\description{
Computes a multi-QTL model with a list QTL candidates (\code{QTL}) and return
the decomposed QTL genetic effects per cross or per parents. The list of QTL
can be of different types (cross-specific, parental, ancestral or bi-allelic).
The type of QTL effects are specified in the vector \code{Q.eff}.
}
\details{
This function computes for each QTL position the genetic effects of the
cross, parental, ancestral or SNP allele components. For the cross-specific
model (\code{Q.eff = "cr"}), the genetics effects represent the substitution
effect of an single allele from the parent 2 (or B) with respect to an allele
coming from the parent 1 or A. All effects are given in absolute value with
the parent that cary the positive allele.

For the parental and the ancestral model (\code{Q.eff = "par" or "anc"}), one
parental (ancestral) allele is set as reference per connected part of the
design. Effects of the other alleles are estimated as deviation with respect
to the reference. For more details about reference definition see
\code{\link{QTL_genEffects}} and \code{\link{design_connectedness}} .

The user can specify if the allele effect that should be put as reference
within connected parts are the most or less used one setting argument
\code{ref.all.most = TRUE or FALSE}. Allele usage is first defined
in term of number of crosses where the allele segregates. Then, if two
alleles segregate in an equal number of crosses, we look at the total cross
sizes. For example in a NAM design the central parent is the most used allele.

For the bi-allelic model (\code{Q.eff = "biall"}), the genetic effects
represent the effects of a single allele copy of the most frequent allele.

\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.
}
\examples{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)

mppData <- USNAM_mppData
mppData_bi <- USNAM_mppData_bi
par.clu <- USNAM_parClu


# Equalize the list of markers of the two mppData objects and par.clu

com.mk.list <- intersect(mppData$map$mk.names, mppData_bi$map$mk.names)

mppData <- mppData_subset(mppData = mppData, mk.list = com.mk.list)
mppData_bi <- mppData_subset(mppData = mppData_bi, mk.list = com.mk.list)
par.clu <- par.clu[rownames(par.clu) \%in\% com.mk.list, ]


QTL.res <- MQE_forward(mppData = mppData, mppData_bi = mppData_bi,
                   Q.eff = c("par", "anc", "biall"), par.clu = par.clu)

QTL.eff <- MQE_genEffects(mppData = mppData, mppData_bi = mppData_bi,
                          QTL = QTL.res[, 1], Q.eff = QTL.res[, 5],
                          par.clu = par.clu)

}
\author{
Vincent Garin
}
\seealso{
\code{\link{mppData_form}}, \code{\link{mpp_SIM}},
\code{\link{parent_cluster}}, \code{\link{USNAM_parClu}}
}

