% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQE_genEffects.R
\name{MQE_genEffects}
\alias{MQE_genEffects}
\title{QTL genetic effects multi-QTL effect model}
\usage{
MQE_genEffects(mppData = NULL, mppData_bi = NULL, QTL = NULL, Q.eff,
  par.clu = NULL, VCOV = "h.err", par.ref = NULL)
}
\arguments{
\item{mppData}{An IBD object of class \code{mppData}
See \code{\link{mppData_form}} for details. Default = NULL.}

\item{mppData_bi}{Required IBS object of class \code{mppData} if the user
wants to allow QTLs with a bi-allelic effect. \strong{The list of marker must
be strictly the same as the one of \code{mppData}.} Default = NULL.}

\item{QTL}{Vector of \code{character} markers or inbetween marker positions
names. Default = NULL.}

\item{Q.eff}{\code{Character} vector indicating for each QTL position the
type of QTL effect among: "cr", "par", "anc" and "biall". For details look at
\code{\link{mpp_SIM}}.}

\item{par.clu}{Required argument if the user wants to allow QTLs with an
ancestral effect. \code{interger matrix} representing the results of a parents
genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{par.ref}{\code{Character} expression indicating the parent or the
ancestral class containing the specified parent that will be set as reference
for the computation of the genetic effect. By default
the function will use the first parent of the \code{mppData$parents} list.
Default = NULL.}
}
\value{
Return:

\item{results}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects per cross or parent.}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}
\item{For cross-specific model, parent with the positive additive effects.}
\item{Allele scores of the parents if \code{geno.par} is non NULL
in the \code{mppData} object.}

  }

}
}
\description{
Computes a multi-QTL model with a list QTL candidates (\code{QTL}) and return
the decomposed QTL genetic effects per cross or per parents. The list of QTL
can be of different types (cross-specific, parental, ancestral or bi-allelic).
The type of QTL effects are specified in the vector \code{Q.eff}.
}
\details{
This function computes for each QTL position the genetic effects of the
cross, parental, ancestral or SNP allele components. For the cross-specific
model (\code{Q.eff = "cr"}), the genetics effects represent the substitution
effect of an single allele from the parent 2 (or B) with respect to an allele
coming from the parent 1 or A. All effects are given in absolute value with
the parent that cary the positive allele.

For the parental or ancestral models (\code{Q.eff = "par"
or "anc"}), the genetic effects represent the effect of a single parental
(ancestral) allele with respect to a reference allelele. The reference
parental (ancestral) allele can be specified in the argument
\code{par.ref}. For the ancestral model, the ancestral class inferred
for the specified parent will be used as reference. For the parental and
ancestral model computed with the homogeneous residual term assumption
(\code{VCOV = "h.err"}), it is also possible to estimate one effect for
each parental (ancestral) allele with the constraint that these effects
sum to zero. This can be done using (\code{const = "sum.0"}).

For the bi-allelic model (\code{Q.eff = "biall"}), the genetic effects
represent the effects of a single allele copy of the most frequent allele.

\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.
}
\examples{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)

mppData <- USNAM_mppData
mppData_bi <- USNAM_mppData_bi
par.clu <- USNAM_parClu


# Equalize the list of markers of the two mppData objects and par.clu

com.mk.list <- intersect(mppData$map$mk.names, mppData_bi$map$mk.names)

mppData <- mppData_subset(mppData = mppData, mk.list = com.mk.list)
mppData_bi <- mppData_subset(mppData = mppData_bi, mk.list = com.mk.list)
par.clu <- par.clu[rownames(par.clu) \%in\% com.mk.list, ]


QTL <- MQE_forward(mppData = mppData, mppData_bi = mppData_bi,
                   Q.eff = c("par", "anc", "biall"), par.clu = par.clu)


QTL.effects <- MQE_genEffects(mppData = mppData, mppData_bi = mppData_bi,
                              QTL = QTL[, 1], Q.eff = QTL[, 5],
                              par.clu = par.clu)

}
\author{
Vincent Garin
}
\seealso{
\code{\link{mppData_form}}, \code{\link{mpp_SIM}},
\code{\link{parent_cluster}}, \code{\link{USNAM_parClu}}
}

