% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQE_proc.R
\name{MQE_proc}
\alias{MQE_proc}
\title{Multi-QTL effect MPP analysis}
\usage{
MQE_proc(pop.name = "MPP_MQE", trait.name = "trait1", mppData = NULL,
  trait = 1, Q.eff, VCOV = "h.err", threshold = 4, window = 30,
  backward = TRUE, alpha.bk = 0.05, plot.MQE = FALSE, n.cores = 1,
  verbose = TRUE, output.loc = getwd())
}
\arguments{
\item{pop.name}{\code{Character} name of the studied population.
Default = "MPP_MQE".}

\item{trait.name}{\code{Character} name of the studied trait.
Default = "trait1".}

\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the \code{mppData} object should be used. Default = 1.}

\item{Q.eff}{\code{Character} vector of possible QTL effects the user want to
test. Elements of \code{Q.eff} can be "cr", "par", "anc" or "biall".
For details look at \code{\link{mpp_SIM}}.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{threshold}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be considered as significant.
Default = 4.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 30.}

\item{backward}{\code{Logical} value. If \code{backward = TRUE},
the function performs
a backward elimination on the list of selected QTLs. Default = TRUE.}

\item{alpha.bk}{\code{Numeric} value indicating the significance level for
the backward elimination. Default = 0.05.}

\item{plot.MQE}{\code{Logical} value. If \code{plot.MQE = TRUE},
the function will plot the last run of the MQE model determination.
Default = FALSE.}

\item{n.cores}{\code{Numeric}. Specify here the number of cores you like to
use. Default = 1.}

\item{verbose}{\code{Logical} value indicating if the steps of MQE_proc should
be printed. It will not affect the printing of the other functions called by
\code{MQE_proc()}, especially the printing of \code{asreml()}.
Default = TRUE.}

\item{output.loc}{Path where a folder will be created to save the results.
By default the function uses the current working directory.}
}
\value{
Return:

\code{List} containing the following items:

\item{n.QTL}{Number of detected QTLs.}

\item{QTL}{\code{Data.frame} with QTL positions.}

\item{R2}{\code{list} containing R squared statistics of the QTL effects.
for details see \code{\link{QTL_R2}}.}

\item{QTL.effects}{\code{List} of genetic effects per QTL.}


Some output files are also saved at the location specified
(\code{output.loc}):

\enumerate{

\item{A QTL report (QTL_REPORT.txt) with: 1) the number of detected QTLs;
2) the global R squared statistics; 3) for each QTL, position information
 and estimated QTL genetic effect per cross or parents.}

\item{The list of QTLs (QTL.txt).}

\item{The QTL R squared statistics (QTL_R2.txt) (for details see
 \code{\link{QTL_R2}}).}

\item{General results of the QTL detection process: Number of QTL and
global adjusted and non-adjusted R squared statistics. (QTL_genResults.txt).}

\item{if \code{plot.MQE = TRUE}, a plot of the last QTL detection run profile
(plot_MQE.pdf).}


}
}
\description{
Build multi-QTL effects (MQE) models in which different QTL effects
(cross-specific, parental, ancestral or bi-allelic) can be assumed at
different loci.
}
\details{
The possible QTL effect that the user wants to allow must be
specified in \code{Q.eff}. The procedure is the following:

\enumerate{

\item{Forward regression to determine a MQE model with different
possible assumptions for the QTL effect at different loci. The function
use.}

\item{Optional backward elimination (\code{backward = TRUE}) on the final
list of detected QTLs.}

\item{Estimation of the QTL genetic effects and R squared statistics.}

\item{Optional plot (\code{plot.MQE = TRUE}) of the last CIM run of the
forward regression using the function.}

}

\strong{WARNING!(1)} The computation of \code{MQE_proc()} function using mixed
models (all models with \code{VCOV} different than \code{"h.err"})
is technically possible but can be irrealistic
in practice due to a reduced computer power. Since a mixed model is computed at
each single position it can take a lot of time. From our estimation it can take
between 20 to 50 times more time than for linear models. If the number of
detected QTL is supposed to be small (until 5) it could still be feasible.

\strong{WARNING!(2)} The computation of random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can sometimes fail. This could be
due to singularities due to a strong correlation between the QTL term(s) and 
the polygenic term. This situation can appear in the parental model.
the error can also sometimes come from the \code{asreml()} function. From
our experience, in that case, trying to re-run the function one or two times
allow to obtain a result.
}
\examples{

\dontrun{

data(mppData)

# Specify a location where your results will be saved
my.loc <- "C:/.../..."

MQE <- MQE_proc(pop.name = "USNAM", trait.name = "ULA", mppData = mppData,
                Q.eff = c("par", "anc", "biall"), output.loc = my.loc)
                 
}


}
\seealso{
\code{\link{mpp_perm}}, \code{\link{mpp_SIM}},
\code{\link{QTL_R2}}
}
\author{
Vincent Garin
}
