% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_clusterCheck.R
\name{parent_clusterCheck}
\alias{parent_clusterCheck}
\title{Check the result of parent clustering}
\usage{
parent_clusterCheck(par.clu)
}
\arguments{
\item{par.clu}{\code{Interger matrix} representing the results of a parents
genotypes clustering. The columns represent the parental lines and the rows
the different markers or in between positions. At a particular
position, parents with the same value are assumed to inherit from the same
ancestor.}
}
\value{
Return:

\code{List} containing the two following objects:

\item{par.clu}{Reorganised \code{par.clu} object with monomorphic position
put back to the parental situation (1 allele per parent).}

\item{marker.prob}{Vector of marker(s) or in between positions where all
parents where groupped into a single group}
}
\description{
Detect marker or position for which all parents have been clustered into
the same ancestral group (monomorphic). Rewrite these positions with one
allele per parent.
}
\examples{


}
\author{
Vincent Garin
}
