% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CV.R
\name{plot_CV}
\alias{plot_CV}
\title{Plot cross-validation QTL profiles}
\usage{
plot_CV(CV.res, main = "CV QTL profiles")
}
\arguments{
\item{CV.res}{Object returned by function \code{\link{mpp_CV}}.}

\item{main}{Title of the graph. Default = "CV QTL profiles".}
}
\description{
Plot cross-validation (CV) QTL profiles using transparency. The
black straight lines are proportional to the number of time a position has
been detected during the entire CV process. Transparency plot are based on
an idea by Pieter Jongsma (see reference).
}
\examples{

\dontrun{

data(mppData)

my.loc <- "C:/..."

CV <- mpp_CV(mppData = mppData, her = .5, output.loc = my.loc, Rep = 3)

plot_CV(CV.res = CV)

}


}
\references{
https://github.com/pieterjongsma/RProfilePlot
}
\seealso{
\code{\link{mpp_CV}}
}
\author{
Vincent Garin
}
