% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genEffects.R
\name{plot_genEffects}
\alias{plot_genEffects}
\title{plot of QTL genetic effects}
\usage{
plot_genEffects(mppData, Qprof, Q.eff, QTL = NULL,
  main = "QTL genetic effects plot", text.size = 18)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{Qprof}{Object of class \code{QTLprof} returned by the function
\code{\link{mpp_SIM}} or \code{\link{mpp_CIM}} calculated with argument
\code{plot.gen.eff = TRUE}.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental
effects; 3) "anc" for ancestral effects.}

\item{QTL}{Optional argument. Object of class \code{QTLlist} representing a
list of selected position obtained with the function \code{\link{QTL_select}}
or vector of \code{character} marker or in between marker positions names.
These positions will be plotted on the graph. Default = NULL.}

\item{main}{Title of the graph. Default =  "QTL genetic effects plot".}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 18.}
}
\description{
Visualisation of the genome-wide significance of the QTL effect per cross or
per parents using the package ggplot2.
}
\details{
It uses the results of a SIM or CIM profile from a
cross-specific, parental and ancestral model.

Plot the significance of the cross-specific or parental components of the QTL
effects using the results of a SIM or CIM profiles from a cross-specific,
a parental or an ancestral model. White colour means no significant effect.

For a cross-specific QTL profile (\code{Q.eff = "cr"}): Red color means
that the allele coming from parent A(1) increases the phenotypic value and
parent B(2) decreases it and blue that parent A(1) decreases the trait and
parent B(2) increases it.

For a parental (\code{Q.eff = "par"}) or an ancestral model
(\code{Q.eff = "anc"}), the results are given per parents. The significance
of the effect must be interpreted as a deviation with respect to the
reference of each connected part. The reference allele is always defined as
the most frequent one. Red (Blue) colour means a signicative negative
(positive) effect with respect to the reference of the connected part.

The reference parental allele can change at each position according to the
segregation rate. The parent are plotted from the top to the bottom according
to the number of time their allele is set as reference.  Therefore
interpretation of the genetic effect plot should be done with caution.
In that case, the plot should be taken as a rough indication of the signal
distribution.

The colour intensity increase with the significance of the effect.
There are five colour intensities according to the p-value of the QTL
effect: 0.05<p-val<0.01; 0.01<p-val<0.001; 0.001<p-val<0.0001;
0.0001<p-val<0.00001 and p-val< 0.00001.

QTL positions passed to the argument \code{QTL} are also drawn on the graph
using dashed lines.
}
\examples{

data(USNAM_mppData)

SIM <- mpp_SIM(mppData = USNAM_mppData, Q.eff = "cr", plot.gen.eff = TRUE)
QTL <- QTL_select(SIM)

# without QTL positions
plot_genEffects(mppData = USNAM_mppData, Qprof = SIM, Q.eff = "cr") 

# with QTL positions
plot_genEffects(mppData = USNAM_mppData, Qprof = SIM, Q.eff = "cr",
QTL = QTL) 

}
\seealso{
\code{\link{mpp_CIM}}, \code{\link{mpp_SIM}}, \code{\link{QTL_select}}
}
\author{
Vincent Garin
}
