% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QTLprof.R
\name{plot_QTLprof}
\alias{plot_QTLprof}
\title{plot QTL profile}
\usage{
plot_QTLprof(Qprof, QTL = NULL, type = "l", main = "QTL profile",
  threshold = 3)
}
\arguments{
\item{Qprof}{Object of class \code{QTLprof} returned by the function
\code{\link{mpp_SIM}} or \code{\link{mpp_CIM}}.}

\item{QTL}{Optional argument. Object of class \code{QTLlist} representing a
list of selected position obtained with the function \code{\link{QTL_select}}.
These positions will be indicated on the graph. Default = NULL.}

\item{type}{\code{Character} expression indicating the type of plot should be
drawn: "l" for lines , "h" for vertical bar. Default = "l".}

\item{main}{Title of the graph. Default = "QTL profile".}

\item{threshold}{\code{Numeric} QTL significance threshold value draw on
the plot. Default = 3.}
}
\description{
Plots the -log10(p-val) profile of a QTL analysis using the package ggplot2.
The user can pass a list of cofactors or QTL position to the argument
\code{QTL}. These positions will be drawn on the graph using dotted lines.
}
\examples{

data(USNAM_mppData)
data(USNAM_mppData_bi)

SIM <- mpp_SIM(mppData = USNAM_mppData, Q.eff = "cr")
QTL <- QTL_select(SIM)
plot_QTLprof(Qprof = SIM, QTL = QTL)

SIM_biall <- mpp_SIM(mppData = USNAM_mppData_bi, Q.eff = "biall")
plot_QTLprof(Qprof = SIM_biall, type = "h")

}
\seealso{
\code{\link{mpp_SIM}}, \code{\link{mpp_CIM}}, \code{\link{QTL_select}}
}
\author{
Vincent Garin
}
