% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mppData.R
\name{print.summary.mppData}
\alias{print.summary.mppData}
\title{Print summary.mppData object}
\usage{
\method{print}{summary.mppData}(x, ...)
}
\arguments{
\item{x}{object of class \code{summary.mppData}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print summary.mppData object
}
\examples{

data(USNAM_mppData)
sum.mppData <- summary(USNAM_mppData)
print(sum.mppData)

}
