% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_matchGeno.R
\name{QC_matchGeno}
\alias{QC_matchGeno}
\title{Match genotypes in the marker matrix and the phenotype data}
\usage{
QC_matchGeno(mk.mat, pheno)
}
\arguments{
\item{mk.mat}{Marker score \code{matrix} or \code{data.frame} with genotypes
as row and markers as column. \strong{Genotype indicators must be the row
names of the matrix.}}

\item{pheno}{Matrix or data frame representing phenotypic data
\strong{with genotype indicators as row names}.}
}
\value{
Return:

\item{new.mk.mat}{Marker \code{matrix} with only the common genotypes.}

\item{new.pheno}{Phenotypic data of the common genotypes.}

\item{mk.mat.only}{Markers present in the marker matrix but not in
the phenotypic data.}

\item{pheno.only}{Markers present in the phenotypic data but not in
the marker matrix.}
}
\description{
Determine which genotypes are common between the genotype marker matrix
and the phenotypic data. Return the marker matrix and the phenotypic values
that correspond to these genotypes.
}
\examples{

data(USNAM_pheno)
data(USNAM_geno)

# The 6 parents of the crosses were not phenotyped
match <- QC_matchGeno(mk.mat = USNAM_geno, pheno = USNAM_pheno)

# Get the consensus phenotype data and genotype matrix
new.mk.mat <- match$new.mk.mat
new.pheno <- match$new.pheno
rm(match)

}
\seealso{
\code{\link{QC_matchMarker}}
}
\author{
Vincent Garin
}
