% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_genEffects.R
\name{QTL_genEffects}
\alias{QTL_genEffects}
\title{QTL genetic effects}
\usage{
QTL_genEffects(mppData, QTL = NULL, Q.eff = "cr", par.clu = NULL,
  VCOV = "h.err", ref.all.most = TRUE)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.
See \code{\link{mppData_form}} for details.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected position obtained with the function \code{\link{QTL_select}} or
vector of \code{character} marker or inbetween marker positions names.
Default = NULL.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. For more details see \code{\link{mpp_SIM}}. Default = "cr".}

\item{par.clu}{Required argument for the ancesral model \code{(Q.eff = "anc")}.
\code{interger matrix} representing the results of a parents genotypes
clustering. The columns represent the parental lines and the rows
the different markers or in between positions. \strong{The columns names must
be the same as the parents list of the mppData object. The rownames must be
the same as the map marker list of the mppData object.} At a particular
position, parents with the same value are assumed to inherit from the same
ancestor. for more details, see \code{\link{USNAM_parClu}} and
\code{\link{parent_cluster}}. Default = NULL.}

\item{VCOV}{\code{Character} expression defining the type of variance
covariance structure used: 1) "h.err" for an homogeneous variance residual term
(HRT) linear model; 2) "h.err.as" for a HRT model fitted by REML using
\code{ASReml-R}; 3) "cr.err" for a cross-specific variance residual terms
(CSRT) model; 4) "pedigree" for a random pedigree term and HRT model;
and 5) "ped_cr.err" for random pedigree and CSRT model.
For more details see \code{\link{mpp_SIM}}. Default = "h.err".}

\item{ref.all.most}{\code{Logical} value specifying which parental
(ancestral) allele should be used as reference. If
\code{ref.all.most = TRUE}, within each connected part of the design,
the most used allele will be used as reference. Allele usage is first defined
in term of number of crosses where the allele segregates. Then, if two
alleles segregate in an equal number of crosses, we look at the total cross
sizes. If \code{ref.all.most = FALSE}, the less used allele will be used as
reference within each connected part. Default = TRUE.}
}
\value{
Return:

\item{results}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects per cross or parent.}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}
\item{For cross-specific model, parent with the positive additive effects.}
\item{For parental and ancestral model, indicator of connected part of the
design and reference.}
\item{Allele scores of the parents if \code{geno.par} is non NULL
in the \code{mppData} object.}

}

}
}
\description{
Computes a multi-QTL model with a list QTL candidates(\code{QTL}) and return
the decomposed QTL effects per cross or per parents.
}
\details{
This function computes for each QTL position the genetic effects of the
cross, parental, ancestral or SNP allele components. For the cross-specific
model (\code{Q.eff = "cr"}), the genetics effects represent the substitution
effect of an single allele from the parent 2 (or B) with respect to an allele
coming from the parent 1 or A. All effects are given in absolute value with
the parent that cary the positive allele.

For the parental and the ancestral model (\code{Q.eff = "par" or "anc"}), it
is possible to estimate maximum n-1 parental or ancestral alleles per connected
part of the design. Connected parts of the design can be determined using
Weeks and Williams (1964) method (\code{\link{design_connectedness}}).
Connected parts of the design are indicated in the results and parental
(ancestral) allele score getting 0 value are the one that could not be
estimated due to singularities. Estimated effects of the other alleles should
be interpreted within connected part as deviation with respect to the
non-estimable effect(s).

The user can specify if the allele effect that should be put as reference
within connected parts are the most or less used one setting argument
\code{ref.all.most = TRUE or FALSE}. Allele usage is first defined
in term of number of crosses where the allele segregates. Then, if two
alleles segregate in an equal number of crosses, we look at the total cross
sizes. For example in a NAM design the central parent is the most used allele.

For the bi-allelic model (\code{Q.eff = "biall"}), the genetic effects
represent the effects of a single allele copy of the most frequent allele.

\strong{WARNING!} The estimation of the random pedigree models
(\code{VCOV = "pedigree" and "ped_cr.err"}) can be unstable. Sometimes the
\code{asreml()} function fails to produce a results and returns the following
message: \strong{\code{GIV matrix not positive definite: Singular pivots}}.
So far we were not able to identify the reason of this problem and to
reproduce this error because it seems to happen randomly. From our
experience, trying to re-run the function one or two times should allow
to obtain a result.
}
\examples{

data(USNAM_mppData)
data(USNAM_mppData_bi)
data(USNAM_parClu)
par.clu <- USNAM_parClu

# QTL candidates

SIM <- mpp_SIM(USNAM_mppData)
QTL <- QTL_select(SIM)

# cross-specific model

QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "cr")
QTL.effects

# parental model

QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL,
                               Q.eff = "par")
QTL.effects

# ancestral model

QTL.effects <- QTL_genEffects(mppData = USNAM_mppData, QTL = QTL, Q.eff = "anc",
                              par.clu = par.clu)
QTL.effects

# bi-allelic model

SIM <- mpp_SIM(USNAM_mppData_bi, Q.eff = "biall")
QTL <- QTL_select(SIM)


QTL.effects <- QTL_genEffects(mppData = USNAM_mppData_bi, QTL = QTL,
Q.eff = "biall")

QTL.effects

}
\author{
Vincent Garin
}
\references{
Weeks, D. L., & Williams, D. R. (1964). A note on the determination of
connectedness in an N-way cross classification. Technometrics, 6(3), 319-324.
}
\seealso{
\code{\link{mppData_form}}, \code{\link{parent_cluster}},
\code{\link{QTL_select}}, \code{\link{USNAM_parClu}}
}

