% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mppData.R
\name{summary.mppData}
\alias{summary.mppData}
\title{Summary of \code{mppData} object}
\usage{
\method{summary}{mppData}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mppData}.}

\item{...}{Ignored.}
}
\description{
\code{summary} for object of class \code{mppData}.
}
\examples{

data(mppData)
summary(mppData)

}
